### 2019-11-08 #1872 @whedon
[PRE REVIEW]: MNE-BIDS: Organizing neurophysiological data into the BIDS format and facilitating their analysis
**Submitting author:** @sappelhoff (<a href="http://orcid.org/0000-0001-8002-0877">Stefan Appelhoff</a>)
**Repository:** <a href="https://github.com/mne-tools/mne-bids" target ="_blank">https://github.com/mne-tools/mne-bids</a>
**Version:** v0.3
**Reviewers:** jdkent, @TomDonoghue
