### 2019-11-08 #1872 @arokem
👋 @tyarkoni: Would you be willing to review this paper for JOSS? (assuming no [COI](https://github.com/openjournals/joss/blob/master/COI.md)? )
