### 2019-11-08 #1872 @arokem
👋 @ftadel :  would you be willing to review this paper for [the Journal of Open Source Software](https://joss.theoj.org/)? The review process is explained [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
