### 2019-11-09 #1872 @jdkent
I would be happy to!

pending...
I'm on the same [nipype zenodo record](https://zenodo.org/record/3476537#.XcZIjd-YU8o) as Alexandre Gramfort, but we did not directly collaborate. Also I depend on some [MNE code in NiBetaSeries](https://github.com/HBClab/NiBetaSeries/blob/a756e85cab21e5e94cb37ebb6ecea3c698330e0c/src/nibetaseries/interfaces/nilearn.py#L43).

I do not believe these minor perceived conflicts should impact my ability to be an impartial reviewer, but I will leave it to @arokem to decide whether the conflict(s) should be waived.
