### 2019-11-09 #1872 @arokem
Thanks! The nipype zenodo record seems fine to me. And depending on MNE seems to make you an appropriate reviewer here. I'll add you, but stand by until I find a second reviewer, before I kick off the review process.
