### 2019-11-09 #1872 @ftadel
I'd be very interested in reviewing this article, but I'll be traveling for a while and I can't guarantee I'd work on it immediately (probably by the end of November, and for sure before mid-December).
If you're not in a rush, you can count me in.
