### 2019-11-11 5b75438
added min_val and max_val finding functions
