### 2019-11-11 be71bc0
added min_val and max_val finding functions
