+++ mne_bids/tests/test_write.py
@@ -553,10 +553,7 @@ def test_edf(_bids_validate):
     data_path = op.join(testing.data_path(), 'EDF')
     raw_fname = op.join(data_path, 'test_reduced.edf')
 
-    raw = mne.io.read_raw_edf(raw_fname, preload=True)
-    # XXX: hack that should be fixed later. Annotation reading is
-    # broken for this file with preload=False and read_annotations_edf
-    raw.preload = False
+    raw = mne.io.read_raw_edf(raw_fname)
 
     raw.rename_channels({raw.info['ch_names'][0]: 'EOG'})
     raw.info['chs'][0]['coil_type'] = FIFF.FIFFV_COIL_EEG_BIPOLAR
@@ -600,8 +597,7 @@ def test_edf(_bids_validate):
     _bids_validate(output_path)
 
     # test anonymize and convert
-    raw = mne.io.read_raw_edf(raw_fname, preload=True)
-    raw.preload = False
+    raw = mne.io.read_raw_edf(raw_fname)
     output_path = _test_anonymize(raw, bids_basename)
     _bids_validate(output_path)
 
