### 2019-11-11 f6f153e
Update examples/convert_group_studies.py
Co-Authored-By: Mainak Jas <jasmainak@users.noreply.github.com>
