+++ mne_bids/tests/test_write.py
@@ -31,9 +31,8 @@ from mne.io.kit.kit import get_kit_info
 
 from mne_bids import (write_raw_bids, read_raw_bids, make_bids_basename,
                       make_bids_folders, write_anat,
-                      get_anonymization_daysback,
-                      get_group_anonymization_daysback)
-from mne_bids.write import _stamp_to_dt
+                      get_anonymization_daysback)
+from mne_bids.write import _stamp_to_dt, _get_anonymization_daysback
 from mne_bids.tsv_handler import _from_tsv, _to_tsv
 from mne_bids.utils import _find_matching_sidecar
 from mne_bids.pick import coil_type
@@ -109,23 +108,25 @@ def test_stamp_to_dt():
 
 
 def test_get_anonymization_daysback():
+    """Test daysback querying for anonymization."""
     data_path = testing.data_path()
     raw_fname = op.join(data_path, 'MEG', 'sample',
                         'sample_audvis_trunc_raw.fif')
     raw = mne.io.read_raw_fif(raw_fname)
-    min_val, max_val = get_anonymization_daysback(raw)
+    daysback_min, daysback_max = _get_anonymization_daysback(raw)
     # max_val off by 1 on Windows for some reason
-    assert abs(min_val - 28461) < 2 and abs(max_val - 36880) < 2
+    assert abs(daysback_min - 28461) < 2 and abs(daysback_max - 36880) < 2
     raw2 = raw.copy()
     raw2.info['meas_date'] = (np.int32(1158942080), np.int32(720100))
     raw3 = raw.copy()
     raw3.info['meas_date'] = (np.int32(914992080), np.int32(720100))
-    min_val, max_val = get_group_anonymization_daysback([raw, raw2, raw3])
-    assert abs(min_val - 29850) < 2 and abs(max_val - 35446) < 2
+    daysback_min, daysback_max = get_anonymization_daysback([raw, raw2, raw3])
+    assert abs(daysback_min - 29850) < 2 and abs(daysback_max - 35446) < 2
     raw4 = raw.copy()
     raw4.info['meas_date'] = (np.int32(4992080), np.int32(720100))
     with pytest.raises(ValueError, match='The dataset spans more time'):
-        min_val, max_val = get_group_anonymization_daysback([raw, raw2, raw4])
+        daysback_min, daysback_max = \
+            get_anonymization_daysback([raw, raw2, raw4])
 
 
 @requires_version('pybv', '0.2.0')
