### 2019-11-12 23896bb
a few last formatting changes
