### 2019-11-12 f821a29
a few last formatting changes
