+++ examples/convert_group_studies.py
@@ -25,8 +25,8 @@ References
 import os.path as op
 
 import mne
-from mne_bids import (write_raw_bids, make_bids_basename,
-                      get_anonymization_daysback)
+from mne_bids import write_raw_bids, make_bids_basename
+from mne_bids.write import _get_anonymization_daysback
 from mne_bids.datasets import fetch_faces_data
 from mne_bids.utils import print_dir_tree
 
@@ -62,30 +62,35 @@ event_id = {
 ###############################################################################
 # Let us loop over the subjects and create BIDS-compatible folder
 
-daysback_min, daysback_max = get_anonymization_daysback(
-    [mne.io.read_raw_fif(op.join(data_path, repo, 'sub%03d' % subject_id,
-                                 'MEG', 'run_%02d_raw.fif' % run))
-     for run in runs for subject_id in subject_ids])
-
+# Get a list of the raw objects for this dataset to use their dates
+# to determine the number of daysback to use to anonymize.
+# While we're looping through the files, also generate the
+# BIDS-compatible names that will be used to save the files in BIDS.
+raw_list = list()
+bids_list = list()
 for subject_id in subject_ids:
     subject = 'sub%03d' % subject_id
     for run in runs:
         raw_fname = op.join(data_path, repo, subject, 'MEG',
                             'run_%02d_raw.fif' % run)
-
         raw = mne.io.read_raw_fif(raw_fname)
+        raw_list.append(raw)
         bids_basename = make_bids_basename(subject=str(subject_id),
                                            session='01', task='VisualFaces',
                                            run=str(run))
+        bids_list.append(bids_basename)
 
-        # By using the same anonymization `daysback` number we can
-        # preserve the longitudinal structure of multiple sessions for a
-        # single subject and the relation between subjects. Be sure to
-        # change or delete this number before putting code online, you
-        # wouldn't want to inadvertently de-anonymize your data.
-        write_raw_bids(raw, bids_basename, output_path, event_id=event_id,
-                       anonymize=dict(daysback=daysback_min + 2117),
-                       overwrite=True)
+daysback_min, daysback_max = _get_anonymization_daysback(raw_list)
+
+for raw, bids_basename in zip(raw_list, bids_list):
+    # By using the same anonymization `daysback` number we can
+    # preserve the longitudinal structure of multiple sessions for a
+    # single subject and the relation between subjects. Be sure to
+    # change or delete this number before putting code online, you
+    # wouldn't want to inadvertently de-anonymize your data.
+    write_raw_bids(raw, bids_basename, output_path, event_id=event_id,
+                   anonymize=dict(daysback=daysback_min + 2117),
+                   overwrite=True)
 
 ###############################################################################
 # Now let's see the structure of the BIDS folder we created.
