### 2019-11-12 3a16700
back to public function
