+++ mne_bids/__init__.py
@@ -4,5 +4,6 @@ __version__ = '0.4.dev0'
 
 
 from mne_bids.write import (write_raw_bids, make_bids_folders, make_bids_basename,  # noqa: E501 F401
-                            make_dataset_description, write_anat)  # noqa: F401
+                            make_dataset_description, write_anat,  # noqa: F401
+                            get_anonymization_daysback)  # noqa: F401
 from mne_bids.read import read_raw_bids, get_head_mri_trans  # noqa: F401
