### 2019-11-12 f5950af
good to go waiting on mne-python #7016 to merge
