### 2019-11-12 2baf8a7
good to go waiting on mne-python #7016 to merge
