### 2019-11-13 da86802
Update mne_bids/write.py
Co-Authored-By: Mainak Jas <jasmainak@users.noreply.github.com>
