### 2019-11-13 51dea88
dealt with None meas_dates
