### 2019-11-13 7240df4
dealt with None meas_dates
