### 2019-11-13 6fcda8e
Luke PR fixes
