### 2019-11-13 a075b35
changed logic so if dates are None, meas_info need not be provided
