+++ mne_bids/write.py
@@ -1111,7 +1111,6 @@ def write_raw_bids(raw, bids_basename, output_path, events_data=None,
 
     # Anonymize
     if anonymize is not None:
-<<<<<<< HEAD
         # if info['meas_date'] None, then the dates are not stored
         if raw.info['meas_date'] is None:
             daysback = None
@@ -1130,22 +1129,6 @@ def write_raw_bids(raw, bids_basename, output_path, events_data=None,
                 raise ValueError('`daysback` exceeds maximum value MNE '
                                  'is able to store in FIF format, must '
                                  'be less than %i' % daysback_max)
-=======
-        if 'daysback' not in anonymize or anonymize['daysback'] is None:
-            raise ValueError('`daysback` argument required to anonymize.')
-        daysback = anonymize['daysback']
-        daysback_min, daysback_max = _get_anonymization_daysback(raw)
-        if daysback < daysback_min:
-            warn('`daysback` is too small; the measurement date '
-                 'is after 1925, which is not recommended by BIDS.'
-                 'The minimum `daysback` value for changing the measurement'
-                 'date of this data to before this date is %i' % daysback_min)
-        if (ext == '.fif' or (kind == 'meg' and convert) and
-                daysback > daysback_max):
-            raise ValueError('`daysback` exceeds maximum value MNE is able '
-                             'to store in FIF format, must be less than %i' %
-                             daysback_max)
->>>>>>> moved convert up for anonymization daysback not exceeding max
         keep_his = anonymize['keep_his'] if 'keep_his' in anonymize else False
         raw.info = anonymize_info(raw.info, daysback=daysback,
                                   keep_his=keep_his)
