### 2019-11-14 09101bc
Update mne_bids/tests/test_write.py
Co-Authored-By: Mainak Jas <jasmainak@users.noreply.github.com>
