### 2019-11-14 0741ea0
cleaning up new tests
