### 2019-11-14 7c3a364
Update mne_bids/write.py
Co-Authored-By: Mainak Jas <jasmainak@users.noreply.github.com>
