### 2019-11-14 #1872 @arokem
👋 @ixjlyons : would you be willing to review this submission to JOSS?
