### 2019-11-14 9693284
in get_head_mri_trans, we need to allow_maxshield=True
