### 2019-11-14 e1547dd
use neutral param + kwargs in private function
