### 2019-11-14 f7d0637
use neutral param + kwargs in private function
