### 2019-11-15 944ae6a
need to allow_maxshield = True when reading raw in get_head_mri_trans
