### 2019-11-15 5504f19
ENH: add function to retrieve emptyroom
