### 2019-11-16 bb09c09
TST improve testing
