### 2019-11-16 #1872 @arokem
👋 @TomDonoghue : would you be willing to review this submission for JOSS?
