### 2019-11-16 #1872 @TomDonoghue
@arokem - Yes, I'd be happy to review this submission!

I am an MNE user, and a fan of the BIDS project, but not directly involved in either, so I think I should be clear for COIs. This seems very relevant to the kind of word I do, and I'd be happy to review it.
