### 2019-11-16 1816356
Update doc/whats_new.rst
Co-Authored-By: Mainak Jas <jasmainak@users.noreply.github.com>
