### 2019-11-16 c6466e9
prevent bug for non fif files
