### 2019-11-16 #1896 @whedon
[REVIEW]: MNE-BIDS: Organizing neurophysiological data into the BIDS format and facilitating their analysis
**Submitting author:** @sappelhoff (<a href="http://orcid.org/0000-0001-8002-0877">Stefan Appelhoff</a>)
**Repository:** <a href="https://github.com/mne-tools/mne-bids" target ="_blank">https://github.com/mne-tools/mne-bids</a>
**Version:** 0.3.1
