### 2019-11-16 #1872 @arokem
Thanks @TomDonoghue and @jdkent ! Please head over to #1896 for your reviewer check lists. Feel free to post issues on the software repository, linking to the review issue. Feel free to tag me into any issues where you'd like me to weigh in. I'll close this pre-review issue now.
