+++ mne_bids/tests/test_write.py
@@ -207,6 +207,12 @@ def test_fif(_bids_validate):
     output_path2 = _test_convert(raw2, bids_basename, events_fname, event_id)
     _bids_validate(output_path2)
 
+    with pytest.warns(UserWarning, match='not supported by BIDS'):
+        write_raw_bids(raw2, bids_basename, output_path2,
+                       events_data=events_fname,
+                       event_id=event_id, convert=False,
+                       overwrite=True, verbose=True)
+
     os.remove(op.join(output_path, 'test-raw.fif'))
     bids_dir = op.join(output_path, 'sub-%s' % subject_id,
                        'ses-%s' % session_id, 'eeg')
