### 2019-11-18 a2ef00c
working example of fid creation and alignment for ecog w/o dig
