+++ examples/add_fiducials_for_defacing.py
@@ -1,0 +1,125 @@
+"""
+================================
+Add fiducial points for Defacing
+================================
+
+In this example, we will add fiducial points so that mne-bids
+knows where the face is for the defacing anonymization algorithm.
+"""
+
+# Authors: Alex Rockhill <aprockhill206@gmail.com>
+# License: BSD (3-clause)
+
+###############################################################################
+# We are importing everything we need for this example:
+import os
+import os.path as op
+import importlib
+import numpy as np
+import matplotlib.pyplot as plt
+from scipy.io import loadmat
+
+from nilearn.plotting import plot_anat
+
+import mne
+from mne.datasets import sample
+
+from mne_bids import write_anat
+
+###############################################################################
+# Load in the raw data to which the dig montage will be added.
+
+data_path = sample.data_path()
+
+# load in the raw data for the fiducials
+raw_fname = op.join(data_path, 'MEG', 'sample', 'sample_audvis_raw.fif')
+raw = mne.io.read_raw_fif(raw_fname)
+
+# Get Landmarks from the file. 0, 1, and 2 correspond to LPA, NAS, RPA
+# and the 'r' key will provide us with the xyz coordinates
+pos = np.asarray((raw.info['dig'][0]['r'],
+                  raw.info['dig'][1]['r'],
+                  raw.info['dig'][2]['r']))
+del raw
+
+# load in the T1
+t1_mgh_fname = op.join(data_path, 'subjects', 'sample', 'mri', 'T1.mgz')
+
+# load in ECoG data
+mat = loadmat(mne.datasets.misc.data_path() + '/ecog/sample_ecog.mat')
+ch_names = mat['ch_names'].tolist()
+elec = mat['elec']  # electrode positions given in meters
+
+###############################################################################
+# Create unaligned datapoints in a dig montage and then align them to the MRI.
+
+# Now we make a montage stating that the sEEG contacts are in head
+# coordinate system (although they are in MRI). This is compensated
+# by the fact that below we do not specicty a trans file so the Head<->MRI
+# transform is the identity.
+montage = mne.channels.make_dig_montage(ch_pos=dict(zip(ch_names, elec)),
+                                        coord_frame='head',
+                                        lpa=pos[0], nasion=pos[1], rpa=pos[2])
+
+# Since human heads are relatively similar in size, aligning the
+# fiducial points from the example data will work pretty well.
+info = mne.create_info(ch_names, 1000., 'ecog', montage=montage)
+
+# Use random data as a placeholder
+raw = mne.io.RawArray(np.random.random((len(ch_names), 1000)), info)
+
+raw.save('sample_ecog.fif')
+
+# Now align the fiducial points and save out the aligned trans file.
+# See https://mne.tools/stable/generated/mne.gui.coregistration.html
+# for instructions on how to do this. Briefly, click on the LPA
+# checkbox, click on the LPA on the head, then do the same for
+# nasion and RPA. Then click save in the same menu box.
+mne.gui.coregistration(subject='sample',
+                       subjects_dir=op.join(data_path, 'subjects'),
+                       inst='sample_ecog.fif')
+
+os.remove('sample_ecog.fif')
+
+# The changed fiducials were saved as sample-fiducials.fif in this case.
+fids, coord_frame = mne.io.read_fiducials(
+    op.join(data_path, 'subjects', 'sample', 'bem', 'sample-fiducials.fif'))
+
+montage2 = mne.channels.make_dig_montage(ch_pos=dict(zip(ch_names, elec)),
+                                         coord_frame='head',
+                                         lpa=fids[0]['r'],
+                                         nasion=fids[1]['r'],
+                                         rpa=fids[2]['r'])
+
[…]
