+++ doc/whats_new.rst
@@ -22,7 +22,7 @@ Current
 Changelog
 ~~~~~~~~~
 
-- Added example of creating fiducial points for defacing aMRI during the anonymization process, by `Alex Rockhill`_ (`# <https://github.com/mne-tools/mne-bids/pull/>`_)
+- Added example of creating fiducial points for defacing aMRI during the anonymization process, by `Alex Rockhill`_ (`#292 <https://github.com/mne-tools/mne-bids/pull/292>`_)
 - Added option to anonymize by shifting measurement date with `anonymize` parameter, in accordance with BIDS specifications, by `Alex Rockhill`_ (`#280 <https://github.com/mne-tools/mne-bids/pull/280>`_)
 - Added automatic conversion of FIF to BrainVision format with warning for EEG only data and conversion to FIF for meg non-FIF data, by `Alex Rockhill`_ (`#237 <https://github.com/mne-tools/mne-bids/pull/237>`_)
 - Add possibility to pass raw readers parameters (e.g. `allow_maxshield`) to :func:`read_raw_bids` to allow reading BIDS-formatted data before applying maxfilter, by  `Sophie Herbst`_
