### 2019-11-18 c3dee51
a bit of cleaning up still
