### 2019-11-18 3ee3bbe
removed mayavi dependancy
