### 2019-11-18 9c29684
removed mayavi from environment
