+++ environment.yml
@@ -12,7 +12,6 @@ dependencies:
 - matplotlib
 - pip:
   - mne>=0.19.1
-  - mayavi
   - nibabel
   - pybv
   - nilearn
