### 2019-11-18 f700f1d
DOC: address alex comment + example
