### 2019-11-19 b584581
a few testing comments
