### 2019-11-19 ff977bc
a few testing comments
