### 2019-11-21 807d949
DOC: Alex and Stefan's suggestions
Co-Authored-By: Alex Rockhill <aprock@uw.edu>
Co-Authored-By: Stefan Appelhoff <stefan.appelhoff@mailbox.org>
