### 2019-11-21 1597977
Update mne_bids/write.py
Co-Authored-By: Mainak Jas <jasmainak@users.noreply.github.com>
