### 2019-11-21 df80797
okay mostly figured out I think
