### 2019-11-21 a5ddb52
fixed for events_from_annotations
