+++ mne_bids/write.py
@@ -609,6 +609,7 @@ def _write_raw_brainvision(raw, bids_fname):
                           'file to Brainvision format')
     from pybv import write_brainvision
     events, _ = events_from_annotations(raw)
+    events[:, 0] -= raw.first_samp
     meas_date = raw.info['meas_date']
     if meas_date is not None:
         meas_date = _stamp_to_dt(meas_date)
