### 2019-11-21 48d0deb
Update whats_new.rst
