+++ doc/whats_new.rst
@@ -31,7 +31,7 @@ Changelog
 Bug
 ~~~
 
-- Fixed broken event timing broken by conversion to BV in :func:`write_raw_bids`, by `Alex Rockhill`_ (`# <https://github.com/mne-tools/mne-bids/pull/>`_)
+- Fixed broken event timing broken by conversion to BV in :func:`write_raw_bids`, by `Alex Rockhill`_ (`#294 <https://github.com/mne-tools/mne-bids/pull/294>`_)
 
 API
 ~~~
