### 2019-11-25 mne-tools/mne-bids#295 @jdkent
[docs,joss] binder examples do not have mne_bids installed
`mne_bids` should be installed as a [dependency for generating the binder notebooks](https://github.com/mne-tools/mne-bids/blob/159797793ed80a169cd5c26641c8d91b84669b27/doc/conf.py#L130).

currently, `mne_bids` is not available for the binder examples ([example](https://mybinder.org/v2/gh/mne-tools/mne-bids/gh-pages?filepath=notebooks/auto_examples/create_bids_folder.ipynb))
