### 2019-11-25 mne-tools/mne-bids#296 @jdkent
[joss] command-line app mne_bids "command not found"
after installing using the instructions, I tried calling the commandline app `mne_bids`, by typing
```
$ mne_bids
```
which returns
```
/usr/local/miniconda3/envs/mne_test/bin/mne_bids: line 1: CLI for MNE-BIDS.: command not found
```

however, if I type:
```
$ python $(which mne_bids)
```
then I get the expected functionality.

If i add `#!/usr/bin/env python` to the file, the script also works as expected on my linux machine (but this solution will not generalize to windows).

A potential general solution could be moving the `bin` directory into the `mne_bids` directory and using the `console_scripts` as an entry point similar to how [fmriprep does it](https://github.com/poldracklab/fmriprep/blob/543827bc161fb99cc3a7d014761cbc7ef6a303f6/setup.cfg#L84) (you can it specified in `setup.py`; it does not have to be in `setup.cfg`).

(or I could be missing a method that should be using the script)
