### 2019-11-25 mne-tools/mne-bids#297 @jdkent
[joss,refactor] similarity to pybids
Definitely not an immediate issue, but I see some duplication of effort between `mne-bids` and `pybids`.

for example, [get_kinds](https://github.com/mne-tools/mne-bids/blob/159797793ed80a169cd5c26641c8d91b84669b27/mne_bids/utils.py#L32) in `mne-bids` appears similar to [get_entities](https://github.com/bids-standard/pybids/blob/24d2032020fd407e5bcbbbf20943726f48717574/bids/layout/layout.py#L602) in `pybids`.

[make_bids_basename](https://github.com/mne-tools/mne-bids/blob/159797793ed80a169cd5c26641c8d91b84669b27/mne_bids/write.py#L745) in `mne_bids` is also similar to [build_path](https://github.com/bids-standard/pybids/blob/24d2032020fd407e5bcbbbf20943726f48717574/bids/layout/writing.py#L62) in `pybids`.

I understand there is a balance between duplication of effort and bringing in extra dependencies, but I wanted to bring this up and see what your opinions are for synchronizing efforts (this may be for further down the line).
