### 2019-11-25 8e533c0
MAINT: modernize CLI to work on Windows
