+++ mne_bids/commands/__init__.py
@@ -1,0 +1,3 @@
+"""Initialize cli."""
+
+from mne_bids.commands import run # noqa: F401
