### 2019-11-25 7c87ad4
DOC: improve narrative
