### 2019-11-25 bb832d4
DOC: improve narrative
