### 2019-11-25 a8791f6
FIX better handling of no matching empty room
