### 2019-11-25 mne-tools/mne-bids#297 @sappelhoff
Hi @jdkent, thanks for bringing up this discussion point.

As you already hint at in your post, our decision for "duplicating" some pybids functionality was driven by the idea to rely on as few dependencies as possible. In fact, we *did* rely on pybids earlier this year, but then actively tried to reduce this dependency for user convenience (see e.g., https://github.com/mne-tools/mne-bids/issues/250#issue-476515380).

The pybids equivalents to `get_kinds` and `make_bids_basename` both rely on the pybids `layout` and are thus dealing with objects. For the mne-bids needs it has so far sufficed to have the short and simple functions like `get_kinds` and `make_bids_basename`.

However, if we in the future find that we require more and more of these basic bids-handling functions, I would see that as a good reason to go back to a pybids dependency.

Perhaps relevant in this context is also this issue: https://github.com/mne-tools/mne-bids/issues/255

cc @jasmainak @agramfort who were part of the discussion whether to use more pybids, or write the needed parts ourselves.
