### 2019-11-25 mne-tools/mne-bids#297 @agramfort
Indeed I suggest to limit dependencies unless there start to have a lot of
code
duplication. Here it's just 2 short functions so I feel it's simpler in
terms of maintenance.
But yes we should not reinvent the wheel.

>
