### 2019-11-25 mne-tools/mne-bids#297 @jasmainak
indeed, we were aware of the powerful querying functionality that pybids provides and we certainly don't want to reinvent the wheel. Here we want to do things that are specific to MEG/EEG/iEEG processing for BIDS which cannot be non-trivially achieved with other software.
