### 2019-11-25 d0cc36c
DOC: CLI documentation works again + pep8
