+++ mne_bids/__init__.py
@@ -7,4 +7,4 @@ from mne_bids.write import (write_raw_bids, make_bids_folders, make_bids_basenam
                             make_dataset_description, write_anat,  # noqa: F401
                             get_anonymization_daysback)  # noqa: F401
 from mne_bids.read import read_raw_bids, get_head_mri_trans  # noqa: F401
-from mne_bids import commands
+from mne_bids import commands  # noqa: F401
