+++ mne_bids/commands/__init__.py
@@ -1,3 +1,1 @@
 """Initialize cli."""
-
-from mne_bids.commands import run # noqa: F401
