### 2019-11-26 8edbd97
[MRG] MAINT: modernize CLI to work on Windows (#298)
* MAINT: modernize CLI to work on Windows

* DOC: CLI documentation works again + pep8
