### 2019-11-26 764a081
refer to setup.py for binder, instead of additional env.yml
