+++ doc/binder/environment.yml
@@ -1,17 +1,0 @@
-name: mne_bids
-channels:
-- defaults
-dependencies:
-- python<3.7
-- pip=19
-- mkl
-- numpy>=1.14
-- scipy>=0.18.1
-- scikit-learn
-- matplotlib
-- pip:
-  - mne>=0.19.1
-  - mne-bids
-  - nibabel
-  - pybv
-  - nilearn
