### 2019-11-26 mne-tools/mne-bids#297 @jdkent
Thanks for the replies! Sounds like you all are well aware of `pybids` and are monitoring how much duplication is occurring between packages. I'll close this issue.
