### 2019-11-27 5809a82
[BUG] [MRG] Pybv event saving bug and adding tests (#294)
* issues

* okay mostly figured out I think

* fixed flake error

* fixed for events_from_annotations

* Update whats_new.rst

* alex comment

* used events from earlier

* sorry this took so long
