+++ mne_bids/write.py
@@ -19,7 +19,7 @@ from scipy import linalg
 from numpy.testing import assert_array_equal
 from mne.transforms import (_get_trans, apply_trans, get_ras_to_neuromag_trans,
                             rotation, translation)
-from mne import Epochs, events_from_annotations
+from mne import Epochs
 from mne.io.constants import FIFF
 from mne.io.pick import channel_type
 from mne.io import BaseRaw, anonymize_info, read_fiducials
@@ -635,7 +635,7 @@ def _write_raw_fif(raw, bids_fname):
         raw.save(bids_fname, split_naming='neuromag', overwrite=True)
 
 
-def _write_raw_brainvision(raw, bids_fname):
+def _write_raw_brainvision(raw, bids_fname, events):
     """Save out the raw file in BrainVision format.
 
     Parameters
@@ -651,12 +651,21 @@ def _write_raw_brainvision(raw, bids_fname):
         raise ImportError('pybv >=0.2.0 is required for converting ' +
                           'file to Brainvision format')
     from pybv import write_brainvision
-    events, event_id = events_from_annotations(raw)
+    # Subtract raw.first_samp because brainvision marks events starting from
+    # the first available data point and ignores the raw.first_samp
+    if events is not None:
+        events[:, 0] -= raw.first_samp
+        events = events[:, [0, 2]]  # reorder for pybv required order
+    meas_date = raw.info['meas_date']
+    if meas_date is not None:
+        meas_date = _stamp_to_dt(meas_date)
     write_brainvision(raw.get_data(), raw.info['sfreq'],
                       raw.ch_names,
                       op.splitext(op.basename(bids_fname))[0],
-                      op.dirname(bids_fname), events[:, [0, 2]],
-                      resolution=1e-6)
+                      op.dirname(bids_fname),
+                      events=events,
+                      resolution=1e-9,
+                      meas_date=meas_date)
 
 
 def _get_anonymization_daysback(raw):
@@ -1223,7 +1232,7 @@ def write_raw_bids(raw, bids_basename, output_path, events_data=None,
         else:
             if verbose:
                 warn('Converting data files to BrainVision format')
-            _write_raw_brainvision(raw, bids_fname)
+            _write_raw_brainvision(raw, bids_fname, events)
     elif ext == '.fif':
         _write_raw_fif(raw, bids_fname)
     # CTF data is saved and renamed in a directory
