+++ doc/whats_new.rst
@@ -32,6 +32,8 @@ Changelog
 Bug
 ~~~
 
+- Fixed broken event timing broken by conversion to BV in :func:`write_raw_bids`, by `Alex Rockhill`_ (`#294 <https://github.com/mne-tools/mne-bids/pull/294>`_)
+
 API
 ~~~
 
