### 2019-12-02 b075098
Adding updated bids root.
