+++ mne_bids/write.py
@@ -847,7 +847,7 @@ def make_bids_folders(subject, session=None, kind=None, output_path=None,
     """Create a BIDS folder hierarchy.
 
     This creates a hierarchy of folders *within* a BIDS dataset. You should
-    plan to create these folders *inside* the output_path folder of the dataset.
+    plan to create these folders *inside* the bids_root folder of the dataset.
 
     Parameters
     ----------
@@ -859,7 +859,7 @@ def make_bids_folders(subject, session=None, kind=None, output_path=None,
     session : str | None
         The session for a item. Corresponds to "ses".
     output_path : str | None
-        The output_path for the folders to be created. If None, folders will be
+        The bids_root for the folders to be created. If None, folders will be
         created in the current working directory.
     make_dir : bool
         Whether to actually create the folders specified. If False, only a
@@ -881,7 +881,7 @@ def make_bids_folders(subject, session=None, kind=None, output_path=None,
     Examples
     --------
     >>> print(make_bids_folders('sub_01', session='my_session',
-                                kind='meg', output_path='path/to/project',
+                                kind='meg', bids_root='path/to/project',
                                 make_dir=False))  # noqa
     path/to/project/sub-sub_01/ses-my_session/meg
 
@@ -973,7 +973,7 @@ def make_dataset_description(path, name=None, data_license=None,
     _write_json(fname, description, overwrite=True, verbose=verbose)
 
 
-def write_raw_bids(raw, bids_basename, output_path, events_data=None,
+def write_raw_bids(raw, bids_basename, bids_root, events_data=None,
                    event_id=None, anonymize=None,
                    overwrite=False, verbose=True):
     """Walk over a folder of files and create BIDS compatible folder.
@@ -1000,7 +1000,7 @@ def write_raw_bids(raw, bids_basename, output_path, events_data=None,
         generated using make_bids_basename.
         Example: `sub-01_ses-01_task-testing_acq-01_run-01`.
         This will write the following files in the correct subfolder of the
-        output_path::
+        bids_root::
 
             sub-01_ses-01_task-testing_acq-01_run-01_meg.fif
             sub-01_ses-01_task-testing_acq-01_run-01_meg.json
@@ -1018,7 +1018,7 @@ def write_raw_bids(raw, bids_basename, output_path, events_data=None,
 
         Note that the modality 'meg' is automatically inferred from the raw
         object and extension '.fif' is copied from raw.filenames.
-    output_path : str
+    bids_root : str
         The path of the root of the BIDS compatible folder. The session and
         subject specific folders will be populated automatically by parsing
         bids_basename.
@@ -1132,22 +1132,22 @@ def write_raw_bids(raw, bids_basename, output_path, events_data=None,
                                  "session date.")
 
     data_path = make_bids_folders(subject=subject_id, session=session_id,
-                                  kind=kind, output_path=output_path,
+                                  kind=kind, output_path=bids_root,
                                   overwrite=False, verbose=verbose)
     if session_id is None:
         ses_path = os.sep.join(data_path.split(os.sep)[:-1])
     else:
         ses_path = make_bids_folders(subject=subject_id, session=session_id,
-                                     output_path=output_path, make_dir=False,
+                                     output_path=bids_root, make_dir=False,
                                      overwrite=False, verbose=verbose)
 
     # create filenames
     scans_fname = make_bids_basename(
         subject=subject_id, session=session_id, suffix='scans.tsv',
         prefix=ses_path)
-    participants_tsv_fname = make_bids_basename(prefix=output_path,
+    participants_tsv_fname = make_bids_basename(prefix=bids_root,
[…]
