+++ mne_bids/commands/tests/test_cli.py
@@ -39,7 +39,7 @@ def test_raw_to_bids(tmpdir):
 
     # Should work
     with ArgvSetter(('--subject_id', subject_id, '--task', task, '--raw',
-                     raw_fname, '--output_path', output_path)):
+                     raw_fname, '--bids_root', output_path)):
         mne_bids_raw_to_bids.run()
 
     # Too few input args
