### 2019-12-02 28de49e
Adding updated bids root.
