### 2019-12-02 a95cb6c
Fixing examples to show the bids_root.
