### 2019-12-02 65cad01
Fixing examples to show the bids_root.
