### 2019-12-02 954e402
Fixing commands in the mne_bids examples.
