### 2019-12-02 593cdc4
Fixing additional locations of output_path renaming to bids_root.
