+++ README.rst
@@ -191,7 +191,7 @@ Example:
 
 .. code-block:: bash
 
-  $ mne_bids raw_to_bids --subject_id sub01 --task rest --raw data.edf --output_path new_path
+  $ mne_bids raw_to_bids --subject_id sub01 --task rest --raw data.edf --bids_root new_path
 
 Bug reports
 -----------
