### 2019-12-02 9ef38f4
Update mne_bids/tests/test_write.py
Co-Authored-By: Mainak Jas <jasmainak@users.noreply.github.com>
