### 2019-12-02 mne-tools/mne-bids#304 @TomDonoghue
Comments on documentation
I have been looking through the code and project, due to reviewing the project for JOSS (https://github.com/openjournals/joss-reviews/issues/1896)

Overall, clearly a very useful, and well organized, developed and documented package. Here I have collected notes of all the possible specific and little things that I noticed in the documentation, that might be useful to be pointed out, in case they are things that the developers consider worth changing / fixing. They are mostly stylistic, and I don't consider any of them to relate to necessary changes (in terms of the JOSS review).

### Documentation Page: CLI
- [x] There is inconsistent styling in the documentation of the CLI options. Sometimes there is a leading capital, sometimes not. Inconsistent use of some trailing periods.
	- `--subject_id` is missing a space before the (01...)

### API Documentation
- [x] `copyfile_bti`: Since this is a public facing function, seems weird not to have full documentation (documenting parameters, for example, like is done for `copyfile_eeglab`)
- [x] Inconsistent docstring style in `copyfiles`. Most of the docs use opening capital letter, and closing period, but this is inconsistent in the `copyfiles` file, which mostly doesn't do this. Are inconsistencies throughout other docstrings as well.
- [x] In `make_bids_basename`, it inconsistently refers to "a item", "the item" or "this item". I also don't find it clear what 'item' refers to in this context.
- [x] In `make_dataset_description`, inconsistent use of whether variable types are referred to as `str` or `string`.
- [x] In `write_raw_bids`, in the warning, I think there is an extra word in the second sentence - it says "convert modify", when I think it just wants to say "modify"

### Examples
- [x] In the `create_bids_filder.py` example, it says "We’ll import the relevant functions from the write module", but specifying the 'write' module seems unneeded, and even possible somewhat confusing for the new user, since it isn't specified in the imports.
- [x] In the `convert_group_studies.py` example, there is some kind of formatting error such that the warning message doesn't get rendered correctly
- [x] In the `convert_mri_and_trans.py` example, I feel like it could be a little easier to see things if the `print_dir_tree(output_path)` lines were separated out from the cells which make BIDS files and write files, since this ends up just collecting tons of output together, and it's harder to find and see the directory print outs.
- [x] The `convert_group_studies` is perhaps a little sparse on descriptive text. It could be perhaps more helpful for the novice reader to note things like, for example, that this is a MEG dataset, where the event_ids come from, and what kind of information is important to provide to `make_bids_basename` and `write_raw_bids`.
- [x] Similarly for the `convert_mne_sample.py` example, there is not a lot of narrative text, and it could perhaps use a little more description for those getting started, and/or perhaps more use of headers etc.
