### 2019-12-03 964a5f0
Refactored everything from output_path to bids_root.
