+++ mne_bids/tests/test_write.py
@@ -72,27 +72,27 @@ def _bids_validate():
         if 'VALIDATOR_EXECUTABLE' != 'n/a':
             bids_validator_exe = ['node', exe]
 
-    def _validate(output_path):
-        cmd = bids_validator_exe + [output_path]
+    def _validate(bids_root):
+        cmd = bids_validator_exe + [bids_root]
         run_subprocess(cmd, shell=shell)
 
     return _validate
 
 
 def _test_anonymize(raw, bids_basename, events_fname=None, event_id=None):
-    output_path = _TempDir()
-    write_raw_bids(raw, bids_basename, output_path,
+    bids_root = _TempDir()
+    write_raw_bids(raw, bids_basename, bids_root,
                    events_data=events_fname,
                    event_id=event_id, anonymize=dict(daysback=33000),
                    overwrite=False)
     scans_tsv = make_bids_basename(
         subject=subject_id, session=session_id, suffix='scans.tsv',
-        prefix=op.join(output_path, 'sub-01', 'ses-01'))
+        prefix=op.join(bids_root, 'sub-01', 'ses-01'))
     data = _from_tsv(scans_tsv)
     if data['acq_time'] is not None and data['acq_time'][0] != 'n/a':
         assert datetime.strptime(data['acq_time'][0],
                                  '%Y-%m-%dT%H:%M:%S').year < 1925
-    return output_path
+    return bids_root
 
 
 def test_stamp_to_dt():
@@ -137,7 +137,7 @@ def test_get_anonymization_daysback():
 @requires_version('pybv', '0.2.0')
 def test_fif(_bids_validate):
     """Test functionality of the write_raw_bids conversion for fif."""
-    output_path = _TempDir()
+    bids_root = _TempDir()
     data_path = testing.data_path()
     raw_fname = op.join(data_path, 'MEG', 'sample',
                         'sample_audvis_trunc_raw.fif')
@@ -148,16 +148,16 @@ def test_fif(_bids_validate):
                            'sample_audvis_trunc_raw-eve.fif')
 
     raw = mne.io.read_raw_fif(raw_fname)
-    write_raw_bids(raw, bids_basename, output_path, events_data=events_fname,
+    write_raw_bids(raw, bids_basename, bids_root, events_data=events_fname,
                    event_id=event_id, overwrite=False)
 
     # Read the file back in to check that the data has come through cleanly.
     # Events and bad channel information was read through JSON sidecar files.
     with pytest.raises(TypeError, match="unexpected keyword argument 'foo'"):
-        read_raw_bids(bids_basename + '_meg.fif', output_path,
+        read_raw_bids(bids_basename + '_meg.fif', bids_root,
                       extra_params=dict(foo='bar'))
 
-    raw2 = read_raw_bids(bids_basename + '_meg.fif', output_path,
+    raw2 = read_raw_bids(bids_basename + '_meg.fif', bids_root,
                          extra_params=dict(allow_maxshield=True))
     assert set(raw.info['bads']) == set(raw2.info['bads'])
     events, _ = mne.events_from_annotations(raw2)
@@ -169,19 +169,19 @@ def test_fif(_bids_validate):
     raw.set_channel_types({raw.ch_names[i]: 'misc'
                            for i in
                            mne.pick_types(raw.info, stim=True, meg=False)})
-    output_path = _TempDir()
+    bids_root = _TempDir()
     with pytest.warns(UserWarning, match='No events found or provided.'):
-        write_raw_bids(raw, bids_basename, output_path, overwrite=False)
+        write_raw_bids(raw, bids_basename, bids_root, overwrite=False)
 
-    _bids_validate(output_path)
[…]
