### 2019-12-03 #1896 @TomDonoghue
I'm going to put comments on the JOSS paper itself here:

#### Content Notes

Broadly, the paper looks good, though I feel like it presumes a fairly high level of familiarity / knowledge of the BIDS and MNE worlds, and so I'm not so sure that the paper is as self-sufficient as one might want to introduce the key ideas to a relatively naive reader. I've added a few notes on things that came to mind, and possible suggestions, for doing so.

Examples:
- In the 1st paragraph, it doesn't indicate how BIDS stores the actual data, and only really refers to metadata. Though full description of BIDS is in other papers, I feel like a quick note along the lines that BIDS is an specification of organizational structure (not a file format) in which data files are stored in existing BIDS-approved file formats might be useful.
- In the 2nd paragraph, what seems to be missing is a note / introduction to what MNE python is. It sort of skips past noting, for example, that MNE is a package for processing electrophysiology data, making the integration of standardized data structures and analysis pipelines a useful thing.
- Somewhat relatedly, the paper doesn't explicitly introduce that MNE-BIDS is a Python package (though this is implicit in the mention of Python objects). I would suggest perhaps more explicitly introducing (perhaps in relation to MNE), that MNE-BIDS is a package in the Python ecosystem.
- In the third paragraph, I feel a bit like it jumps into mechanics while somewhat missing description of the goal of doing so. Potentially there could be an introductory sentence to this paragraph, perhaps along the lines of "The key functionality of the MNE-BIDS package is to take a file structure containing an existing dataset of EEG, MEG or iEEG data, and convert this file structure and metadata organization into a BIDS compatible format."
- Also, in the third paragraph, what does it mean to be BIDS compatible? This concept is presumed, but if one isn't already quite familiar with the notion of BIDS, I'm not sure this makes sense (this also relates back to para 1 not fully introducing how BIDS is structured). I *think* this means if file formats are not already a BIDS approved format then this tool can convert them, and if so I feel like this could be explained slightly more.

#### Typo Notes
- 2nd paragraph, second to last line: after word 'preprocessing', space before comma

---
The content notes are meant as possible suggestions, though not reflecting any changes I'd consider strictly necessary.
