### 2019-12-03 mne-tools/mne-bids#305 @TomDonoghue
Project Comments (JOSS Review)
I have been looking through the code and project, due to reviewing the project for JOSS (openjournals/joss-reviews#1896).

Here I am collecting a couple open thoughts / questions about the project:
- [x] Is there a minimum version of Python3 required? I guess that there is some minimum required sub-version of Py3 required, but it's not clear what that would be, and it might be useful to note that.
- [x] Is it not standard to list out the required dependencies in a `requirements.txt` file? It could be this is no longer considered standard practice, but it struck me as odd that this isn't included.
- [x] On the website, the text overlaps for inputs and docstrings. Not sure if there is a sphinx setting to fix this.
- [x] In the `Contributing.md` file, it is noted that the BIDS validator is required to run tests locally. I think it's worth specifying that this requires the CommandLine app (and not, for example, the Python library). It took me a while to get tests running, partly because I had to chase out how to get BIDSValidator set up as needed.
- [x] Related to tests and dependencies, the current dev version fails a series of tests due to missing `pybv` if it's not installed, but this isn't listed as a requirement in the repo info nor in the Contributing file.
- [x] Note: Once I solved BIDSValidator and pybv, I did get the tests running locally, though one still fails - the `test_ctf` check fails on the last error check, for the "All measreument dates are None" part. I'm not sure why - I didn't dig into this.
- [x] It might be worth adding a section to the README which links to the Contributing file, because otherwise this is not linked from the website, nor easily accessed from the README, and not totally obvious to find.
- [x] Perhaps my most general comment, is that if one starts at the examples, I'm not sure it's particularly clear where to start. I wonder if it might be useful for the new user if the examples were clustered into sections, for example into 'Intro', 'EEG', 'MEG' or something. And/or, perhaps, they could be ordered, with the most introductory example positioned first.
- [x] Relatedly, I wonder if it would be a useful extra example / starting point would be a file that explains the BIDS logic, in the context of electrophysiological data. The README links to the general BIDS site, but the general spec is more about MRI, and if one didn't already know, I don't think there's a clear path offered to motivate what BIDS means / looks like for EEG/MEG/iEEG. And 'Intro' example could briefly introduce some ideas, link to the separate specs more explicitly, and note things like supported file formats, etc.
    - [x] Or, another way to put it: the current site presumes one already know about MNE & BIDS and already knows why they want to use them together. In so far as it's considered within scope to motivate and promote why one should use BIDS, and what it looks like (within code / practical examples) I think a little more could be added (though this could reasonable be considered out of scope).
