### 2019-12-03 #1896 @TomDonoghue
Okay, I have done my main review. In my review, I looked through the code and repository, installed the software, and ran some of the examples. Everything looks good to me, and I think this is a really nice tool, and well developed and useful.

In terms of comments on the repository and code, I have made issues on the main repository, with some notes / ideas on general project thoughts (https://github.com/mne-tools/mne-bids/issues/305) and some little comments on style and consistency in the comments (https://github.com/mne-tools/mne-bids/issues/304). Comments on the paper are above.

Nothing came up that I consider fundamentally important in a way that I think this project could not be approved for publication in JOSS based on the posted guidelines, so all the comments are offered basically as any possible and optional suggestions that came up as I worked through the project.

I left a couple things  unchecked, that are point most related to the main comments I brought up in paper comments and issues on the repository itself - I'm happy to check them off once the authors check through and say if they plan to change anything, so we can check back through those things is so.
