+++ README.rst
@@ -61,9 +61,10 @@ Installation
 ------------
 
 We recommend the `Anaconda <https://www.anaconda.com/download/>`_ Python
-distribution. We require that you use Python 3. You may choose to install
-``mne-bids`` `via pip <#Installation via pip>`_ or
-`via conda <#Installation via conda>`_
+distribution. We require that you use Python 3.5 or higher.
+You may choose to install ``mne-bids``
+`via pip <#Installation via pip>`_ or
+`via conda <#Installation via conda>`_.
 
 Installation via pip
 ####################
