### 2019-12-03 mne-tools/mne-bids#305 @sappelhoff
Thanks for your review @TomDonoghue! I edited your post to turn the bullet points into check marks, so that we can mark more easily which points of yours we have already addressed. I hope that was OK with you :-)

Now, to get started:

> Is there a minimum version of Python3 required? I guess that there is some minimum required sub-version of Py3 required, but it's not clear what that would be, and it might be useful to note that.

Yes, minimum 3.5, same as for MNE-Python. Thanks for marking this, I have updated it in https://github.com/mne-tools/mne-bids/pull/306/commits/aff8fdc99da075522e23d9911a5e4786d9cd8af6

> Is it not standard to list out the required dependencies in a requirements.txt file? It could be this is no longer considered standard practice, but it struck me as odd that this isn't included.

In our installation we recommend that users work with [Anaconda](https://www.anaconda.com/), which usually works in "environments", so we provide a [`environment.yml`](https://github.com/mne-tools/mne-bids/blob/5809a8204f0fc9f18e772807312ecfd656a6d6d9/environment.yml#L1) file, which we also use in our CIs (see e.g., [here](https://github.com/mne-tools/mne-bids/blob/5809a8204f0fc9f18e772807312ecfd656a6d6d9/.travis.yml#L58)). Note however, that we try to keep installation very basic and assume that users do so by following how it's specified in the [README](https://github.com/mne-tools/mne-bids/blob/master/README.rst) ... I don't think a `requirements.txt` file would significantly simplify the installation at this point.
