### 2019-12-03 8e72542
specify which bids-validator type should be installed
