### 2019-12-03 eaa439b
add section to README referring to contributing guide
