### 2019-12-03 mne-tools/mne-bids#305 @sappelhoff
> In the Contributing.md file, it is noted that the BIDS validator is required to run tests locally. I think it's worth specifying that this requires the CommandLine app

I absolutely agree - thanks! See the fix here: https://github.com/mne-tools/mne-bids/pull/306/commits/2cf52b85b91771630044823b0e0b4d9e75547813

> It might be worth adding a section to the README which links to the Contributing file, because otherwise this is not linked from the website, nor easily accessed from the README, and not totally obvious to find.

That's a fair point. I guess it's a bit of a hidden knowledge (or not enforced standard) that [the contributing guide in GitHub is always to be found in CONTRIBUTING.md](https://help.github.com/en/github/building-a-strong-community/setting-guidelines-for-repository-contributors#adding-a-contributing-file)...  I have addressed your point by adding a section to our README here: https://github.com/mne-tools/mne-bids/pull/306/commits/4d3dbd81a6b0fe8c0a1a3b296fac6fc0c60effdc

Better be safe than losing a potential contributor :-)
