### 2019-12-03 c9caf5a
extend list of pkgs to be installed for contributors
