### 2019-12-03 mne-tools/mne-bids#305 @sappelhoff
>  On the website, the text overlaps for inputs and docstrings. Not sure if there is a sphinx setting to fix this.

I tried to find what you mean by this @TomDonoghue but I didn't see it in either stable or latest docs ... could you please send a screenshot and a link to the page where you notice the issue?

> Related to tests and dependencies, the current dev version fails a series of tests due to missing pybv if it's not installed, but this isn't listed as a requirement in the repo info nor in the Contributing file.

Good catch, thanks. I actually went ahead and added *all* optional packages to the "to be installed" list for contributors: https://github.com/mne-tools/mne-bids/pull/306/commits/63aa343ed006730317cde897bf15c1f4d697239f ... note that normal users (as opposed to contributors) are not expected to run tests, so I only added this to the CONTRIBUTORS guide.

> Relatedly, I wonder if it would be a useful extra example / starting point would be a file that explains the BIDS logic, in the context of electrophysiological data. ...

I agree that users of `mne-bids` need a lot of knowledge about BIDS (and MNE-Python) already, but I do think that it's out of the scope of our package to provide a large introduction. Better places for this are the [bids starter kit](https://github.com/bids-standard/bids-starter-kit), however I am interested to hear @jasmainak's opinion on this.

To think practical, I can imagine that we add a sentence or two to our README to better point users towards the MEEG aspect of BIDS (instead of only to BIDS in very general terms)
