### 2019-12-03 mne-tools/mne-bids#304 @sappelhoff
Thanks again @TomDonoghue - like in #305 I took the liberty to turn your bullet points into check marks, so that we can track which of your points we have addressed and which remain open. Thanks also for splitting your review in terms of code / docs / paper. This is really helpful. I'll try to address the points piece by piece (probably with the help of @jasmainak and others :smile: )
