### 2019-12-03 mne-tools/mne-bids#305 @TomDonoghue
For the docs overlap page: sorry, I wasn't super specific, I meant an issue on the rendering of the API page. However, with a little more checking, I realized I had opened it up on (an old version of?) Safari, in which I see things like this:
But that this doesn't happen on Firefox or Chrome. I think this is more about one weird browser issue, and not something with your settings - sorry for not looking into that more!

For the last point, I do totally agree that really diving into explaining BIDS is out of scope, but I think what I was trying to get at is that it doesn't feel like there is a super clear path offered for a novice user who, say, sees this repo on `mne-tools` and wants to learn about BIDS, in the context of M/EEG. I think, like you suggest, even a quick statement like "If you are interested in learning more about BIDS the main page is here, a starter kit is here, and the specifications for MEG, EEG, iEEG are here, here, and here" in the README might help. In the context of a possible example, I was thinking something that does presume one has read, the EEG BIDS paper or spec, for example, but does a little more demonstration like "here is an example BIDS file structure", "here is an example BIDS metadata file", "here is a list of supported file formats", "here is a list of identifiers that can go in file names", etc. But again - up to y'all how you want to organize these kinds of things across the BIDS ecosystem.

All the other noted points & fixes make sense and look good to me!
