### 2019-12-03 mne-tools/mne-bids#305 @jasmainak
Sorry I am slow to catch up. Thanks @TomDonoghue for all the excellent suggestions for improvements.

I am fine adding some description of what BIDS is -- a standard for organizing data. So someone who knows MNE but not BIDS can become curious.

Also @sappelhoff , perhaps we could [order our examples](https://sphinx-gallery.github.io/stable/configuration.html#sorting-gallery-subsections) so that they follow some logical sequence going from easy introduction to what the file structure looks like, to reading, conversion, anonymizing etc?
