### 2019-12-03 6514c1f
FIX: check meas_date=None with CTF data
