### 2019-12-04 8e517a8
need to constrain versions post-hoc for build to work
