+++ .travis.yml
@@ -6,9 +6,9 @@ language: python
 env:
   global:
     # can be "stable", or anything that can be used with git checkout
-    - VALIDATOR_VERSION="master"
+    - VALIDATOR_VERSION="9d8a6ffae5e0ee117b97bc8c978fe34ec3c13af4"
     # can be any branch of MNE-Python (e.g., "maint/0.18", "master", ...)
-    - MNE_VERSION="master"
+    - MNE_VERSION="maint/0.18"
 
 
 before_install:
