### 2019-12-05 0e375a5
specify mne-bids version
