+++ doc/binder-requirements.txt
@@ -1,1 +1,1 @@
-mne_bids
+mne_bids<0.4
