### 2019-12-05 99bd781
use environment.yml in appveyor
