### 2019-12-05 4ee40a0
try with mne 0.19, and not constraining numpy/scipy too much
