+++ .circleci/config.yml
@@ -31,7 +31,7 @@ jobs:
                   conda update --yes --quiet conda
                   conda env create -f environment.yml --name=testenv
                   source activate testenv
-                  pip install mne==0.18.2
+                  pip install mne==0.19
                   python setup.py develop
 
             - run:
