+++ doc/binder-requirements.txt
@@ -1,1 +1,1 @@
-mne_bids<0.4
+mne_bids<=0.3
