+++ environment.yml
@@ -11,7 +11,7 @@ dependencies:
 - scikit-learn
 - matplotlib
 - pip:
-  - mne=0.19
+  - mne==0.19
   - nibabel
   - nilearn
   - flake8
