### 2019-12-05 5aaa02b
big change in readme
