+++ README.rst
@@ -40,7 +40,7 @@ MNE-BIDS
 ========
 
 
-This is a repository for creating
+This is a repository for creating really cool
 `BIDS <http://bids.neuroimaging.io/>`_\ -compatible datasets with
 `MNE <https://mne.tools/stable/index.html>`_.
 
