### 2019-12-05 20fbb88
[MISC] address issues raised in context of JOSS review (#306)
* specify python minimum version

* specify which bids-validator type should be installed

* add section to README referring to contributing guide

* extend list of pkgs to be installed for contributors

* address mainaks comments
