+++ CONTRIBUTING.md
@@ -21,14 +21,25 @@ pypi. Update your installation as follows.
     $ pip uninstall mne  # uninstall pypi mne
     $ pip install -e .  # use the cloned repo for a local install of mne
 
-Then, install the following python packages:
+Then, install the following python packages for testing purposes:
 
     $ pip install flake8 pytest pytest-cov
+
+Afterwards, also install all optional dependencies of `mne-bids` and their
+dependencies:
+
+    $ pip install scikit-learn nibabel nilearn pybv
 
 Finally, it is necessary to install the
 [BIDS validator](https://github.com/bids-standard/bids-validator). The outputs
 of MNE-BIDS are run through the BIDS validator to check if the conversion
 worked.
+You will need the `command line version` of the validator, to be installed via
+[Node.js](https://nodejs.org/en/).
+
+- For installing the *stable* version of the bids-validator, please follow the
+instructions as detailed in the README of the bids-validator repository.
+- For installing the *development* version, see [here](https://github.com/bids-standard/bids-validator/blob/master/CONTRIBUTING.md#using-the-development-version-of-bids-validator)
 
 ##### Building the documentation
 
