### 2019-12-05 mne-tools/mne-bids#305 @sappelhoff
For the sake of completion:

> Note: Once I solved BIDSValidator and pybv, I did get the tests running locally, though one still fails - the test_ctf check fails on the last error check, for the "All measreument dates are None" part. I'm not sure why - I didn't dig into this.

@jasmainak fixed this in #307

Next up:

1. adding some meat to the description on BIDS
2. re-organizing examples
