### 2019-12-05 c133da9
[PAPER] make clear that BIDS is *not* a file format
