### 2019-12-05 c023d2a
[PAPER] explain BIDS compatibility
