### 2019-12-05 30ecc10
[PAPER] explain MNE-Python, mention we are talking about Python software
