### 2019-12-06 18c2306
bump mne to 0.19.X ...always taking highest available X
