### 2019-12-06 5f9a908
describe BIDS better in README
