### 2019-12-06 3f17375
add explicit order to examples
