### 2019-12-06 mne-tools/mne-bids#305 @sappelhoff
Done:

1. add BIDS description: https://github.com/mne-tools/mne-bids/pull/312/commits/5f9a908ffa6827bcc62e5c61273e49821123b3b6

2. re-organized examples: https://github.com/mne-tools/mne-bids/pull/312/commits/3f17375427598c8a19870f2122e465ce40bf6629 ... loosely ordered from basic to advanced.
