### 2019-12-06 3f1641d
consistent capitalization, fix typos
