+++ README.rst
@@ -43,7 +43,7 @@ This is a repository for creating
 `BIDS <http://bids.neuroimaging.io/>`_\ -compatible datasets with
 `MNE-Python <https://mne.tools/stable/index.html>`_.
 
-BIDS (*B*rain *I*maging *D*ata *S*tructure) is a standard to organize data
+BIDS (Brain Imaging Data Structure) is a standard to organize data
 according to a set of rules that describe ...
 
 - how to name your files
@@ -52,8 +52,8 @@ according to a set of rules that describe ...
 
 MNE-Python is a software package for analyzing neurophysiology data.
 
-MNE-BIDS links BIDS and MNE with the goal to make your analyses faster to code,
-more robust to errors, and easily sharable with colleagues.
+**MNE-BIDS links BIDS and MNE with the goal to make your analyses faster to code,
+more robust to errors, and easily sharable with colleagues.**
 
 Documentation
 -------------
