### 2019-12-06 adf4b3d
fix docstrs in copyfiles
