### 2019-12-06 763b3b8
consistent capitalization and punctuation in copyfiles
