### 2019-12-06 0cd471e
fix make_bids_basename docstr
