+++ mne_bids/write.py
@@ -916,7 +916,7 @@ def make_dataset_description(path, name=None, data_license=None,
         The license under which this datset is published.
     authors : list | str | None
         List of individuals who contributed to the creation/curation of the
-        dataset. Must be a list of strings or a single comma separated string
+        dataset. Must be a list of str or a single comma separated str
         like ['a', 'b', 'c'].
     acknowledgements : list | str | None
         Either a str acknowledging individuals who contributed to the
@@ -927,11 +927,11 @@ def make_dataset_description(path, name=None, data_license=None,
         publications that should be cited.
     funding : list | str | None
         List of sources of funding (e.g., grant numbers). Must be a list of
-        strings or a single comma separated string like ['a', 'b', 'c'].
+        str or a single comma separated str like ['a', 'b', 'c'].
     references_and_links : list | str | None
         List of references to publication that contain information on the
-        dataset, or links.  Must be a list of strings or a single comma
-        separated string like ['a', 'b', 'c'].
+        dataset, or links.  Must be a list of str or a single comma
+        separated str like ['a', 'b', 'c'].
     doi : str | None
         The DOI for the dataset.
 
