### 2019-12-06 20e65c8
fix warning in example
