### 2019-12-06 e1098d9
remove confusing info from example
