### 2019-12-06 9969032
separate printing dir tree from processing output
