### 2019-12-06 440ac45
[MRG] use mne-bids for binder (#308)
* use mne-bids for binder

* need to constrain versions post-hoc for build to work

* specify mne-bids version

* try with mne 0.19, and not constraining numpy/scipy too much

* pip syntax

* leave appveyor alone

* fix appveyor

* bump mne to 0.19.X ...always taking highest available X
