+++ doc/binder-requirements.txt
@@ -1,0 +1,1 @@
+mne_bids<=0.3
