+++ appveyor.yml
@@ -4,9 +4,9 @@ environment:
         # latter case, you MUST adjust VALIDATOR_EXECUTABLE to the following value:
         # "C:\\projects\\mne-bids\\bids-validator\\bids-validator\\bin\\bids-validator"
         # ... whereas for "stable", VALIDATOR_EXECUTABLE MUST be set to "n/a"
-        VALIDATOR_VERSION: "master"
+        VALIDATOR_VERSION: "9d8a6ffae5e0ee117b97bc8c978fe34ec3c13af4"
         # can be any branch of MNE-Python (e.g., "maint/0.18", "master", ...)
-        MNE_VERSION: "master"
+        MNE_VERSION: "maint/0.19"
         VALIDATOR_EXECUTABLE: "C:\\projects\\mne-bids\\bids-validator\\bids-validator\\bin\\bids-validator"
         NODEJS_VERSION: "10.16.1"
         PYTHON: "C:\\conda"
@@ -14,6 +14,8 @@ environment:
       - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2017
         PYTHON_VERSION: "3.6"
         PYTHON_ARCH: "64"
+        PYTHON_ROOT: "C:\\Miniconda36-x64"
+        CONDA_INSTRUMENTATION_ENABLED: "true"
 
 install:
   # https://www.appveyor.com/docs/lang/nodejs-iojs/
@@ -32,11 +34,8 @@ install:
       yarn &&
       cd .. &&
       set PATH="%PATH%;C:\projects\mne-bids\bids-validator\bids-validator\bin\" )
-  - "git clone --depth 1 git://github.com/astropy/ci-helpers.git"
-  - "powershell ci-helpers/appveyor/install-miniconda.ps1"
-  - "SET PATH=%PYTHON%;%PYTHON%\\Scripts;%PATH%"
-  - "activate test"
-  - "conda install numpy scipy"
+  - CALL "%PYTHON_ROOT%\\Scripts\\activate.bat"
+  - "conda install -y python=%PYTHON_VERSION% conda numpy scipy"
   - "pip install nibabel flake8 pytest pytest-sugar pytest-faulthandler"
   - "pip uninstall -yq mne"
   - ps: |-
