+++ _sources/auto_examples/convert_mne_sample.rst.txt
@@ -1,7 +1,7 @@
 .. note::
     :class: sphx-glr-download-link-note
 
-    Click :ref:`here <sphx_glr_download_auto_examples_convert_mne_sample.py>` to download the full example code or run this example in your browser via Binder
+    Click :ref:`here <sphx_glr_download_auto_examples_convert_mne_sample.py>` to download the full example code or to run this example in your browser via Binder
 .. rst-class:: sphx-glr-example-title
 
 .. _sphx_glr_auto_examples_convert_mne_sample.py:
@@ -33,6 +33,7 @@ BIDS-compatible folder.
 
 
 
+
 Let us import mne_bids
 
 
@@ -48,6 +49,7 @@ Let us import mne_bids
 
     from mne_bids import write_raw_bids, read_raw_bids, make_bids_basename
     from mne_bids.utils import print_dir_tree
+
 
 
 
@@ -80,6 +82,7 @@ And define the paths and event_id dictionary.
  .. code-block:: none
 
     Using default location ~/mne_data for sample...
+
 
 
 
@@ -216,6 +219,10 @@ Specify the raw_file and events_data and run the BIDS conversion.
     /home/circleci/project/mne_bids/write.py:407: UserWarning: No line frequency found, defaulting to 50 Hz
       warn('No line frequency found, defaulting to 50 Hz')
     Reading 0 ... 166799  =      0.000 ...   277.714 secs...
+    /home/circleci/miniconda/envs/testenv/lib/python3.6/site-packages/mne/utils/docs.py:830: DeprecationWarning: Function read_montage is deprecated; ``read_montage`` is deprecated and will be removed in v0.20. Please use ``read_dig_fif``, ``read_dig_egi``, ``read_custom_montage``, or ``read_dig_captrack`` to read a digitization based on your needs instead; or ``make_standard_montage`` to create ``DigMontage`` based on template; or ``make_dig_montage`` to create a ``DigMontage`` out of np.arrays
+      warnings.warn(msg, category=DeprecationWarning)
+    /home/circleci/miniconda/envs/testenv/lib/python3.6/site-packages/mne/utils/docs.py:813: DeprecationWarning: Class Montage is deprecated; Montage class is deprecated and will be removed in v0.20. Please use DigMontage instead.
+      warnings.warn(msg, category=DeprecationWarning)
 
     Writing '/home/circleci/mne_data/MNE-sample-data/../MNE-sample-data-bids/sub-01/ses-01/meg/sub-01_ses-01_task-audiovisual_run-01_meg.json'...
 
@@ -252,6 +259,8 @@ Specify the raw_file and events_data and run the BIDS conversion.
     Overwriting existing file.
     Writing /home/circleci/mne_data/MNE-sample-data-bids/sub-01/ses-01/meg/sub-01_ses-01_task-audiovisual_run-01_meg.fif
     Closing /home/circleci/mne_data/MNE-sample-data-bids/sub-01/ses-01/meg/sub-01_ses-01_task-audiovisual_run-01_meg.fif [done]
+
+    '/home/circleci/mne_data/MNE-sample-data/../MNE-sample-data-bids'
 
 
 
@@ -373,6 +382,8 @@ Specify some empty room data and run BIDS conversion on it.
     Writing /home/circleci/mne_data/MNE-sample-data-bids/sub-emptyroom/ses-20021206/meg/sub-emptyroom_ses-20021206_task-noise_meg.fif
     Closing /home/circleci/mne_data/MNE-sample-data-bids/sub-emptyroom/ses-20021206/meg/sub-emptyroom_ses-20021206_task-noise_meg.fif [done]
 
[…]
