### 2019-12-07 mne-tools/mne-bids#315 @jdkent
[joss] add code of conduct
Here are a couple examples for a code of conduct
- [fmriprep](https://github.com/poldracklab/fmriprep/blob/5d0c823921a2418f95ca5ff4ea499569971212b9/CODE_OF_CONDUCT.md)
- [nibetaseries](https://github.com/HBClab/NiBetaSeries/blob/a756e85cab21e5e94cb37ebb6ecea3c698330e0c/CODE_OF_CONDUCT.rst)
- [STEMMRoleModels](https://github.com/KirstieJane/STEMMRoleModels/blob/3510453ac71cfa79a6713ca78bdbe72482d8dc08/CODE_OF_CONDUCT.md)
