### 2019-12-07 mne-tools/mne-bids#316 @jdkent
[joss] add issue_template
Is there any interest in adding an issue template for the repository?
examples:
- [nipype](https://github.com/nipy/nipype/blob/63e5ef25eeaf8717ef271cc3397b41c2e754c5ef/.github/ISSUE_TEMPLATE.md)
- [nibetaseries](https://github.com/HBClab/NiBetaSeries/tree/a756e85cab21e5e94cb37ebb6ecea3c698330e0c/.github/ISSUE_TEMPLATE)

This could help guide users towards giving useful bug reports/feature requests.
