### 2019-12-07 mne-tools/mne-bids#318 @jdkent
[joss] Paper Review
[Here are some suggestions for edits](https://docs.google.com/document/d/1pJHUK5boPN7LqkbEBP42Yxahf4IwbAofVr2qUZuvXkE/edit?usp=sharing) to the paper to help emphasize the great aspects of this project. Great Work!
