### 2019-12-07 #1896 @jdkent
Yes, I had a chance to take a look.
`MNE-BIDS` is a useful package for converting MEG, EEG, and (experimentally) iEEG datasets into BIDS format and interfacing with those BIDS organized datasets.

I have made issues pertaining to the following sections:

## Functionality
- [x] commandline app not working as intended: mne-tools/mne-bids#296 (addressed)
- [x] similarity/differences from pybids: mne-tools/mne-bids#297 (addressed)

## Documentation
- [ ] fix binder examples: mne-tools/mne-bids#295 (ongoing)
- [ ] add code of conduct: mne-tools/mne-bids#315 (open)
- [ ] add issue templates: mne-tools/mne-bids#316 (open)
- [ ] ensure correct python version: mne-tools/mne-bids#314 (open)

## Software Paper
- [ ] suggestings for writing style: mne-tools/mne-bids#318 (open)

MNE-BIDS looks like a great addition to JOSS! Great work to the authors! I left a few comments to improve the documentation and make interactions with the project better. I also offered some suggested edits to the paper to promote clarity/conciseness. As soon as the above issues are addressed I'll check off the final marks in the review above.
