### 2019-12-07 #1896 @arokem
Fantastic! Thank you both for the detailed and valuable reviews.

@sappelhoff: please let me know if you have any questions and/or ping me here when you have had a chance to address the comments.
