### 2019-12-07 a09bd74
use travis build with minimal deps
