### 2019-12-07 72e1bd3
[MRG] use environment.yml in appveyor (#309)
* use environment.yml in appveyor

* fix typo
