### 2019-12-07 74498ea
Apply suggestions from code review
Co-Authored-By: Mainak Jas <jasmainak@users.noreply.github.com>
