+++ examples/create_bids_folder.py
@@ -1,6 +1,6 @@
 """
 ==============================================
-01. Creating BIDS-compatible folders and files
+01. Creating BIDS-compatible folder names and filenames
 ==============================================
 
 The Brain Imaging Data Structure (BIDS) has standard conventions for file
