+++ doc/binder-requirements.txt
@@ -1,1 +1,0 @@
-mne_bids<=0.3
