+++ _sources/auto_examples/convert_mne_sample.rst.txt
@@ -405,27 +405,27 @@ Now let's see the structure of the BIDS folder we created.
 
     |MNE-sample-data-bids/
     |--- participants.tsv
-    |--- dataset_description.json
     |--- participants.json
+    |--- dataset_description.json
     |--- sub-emptyroom/
     |------ ses-20021206/
     |--------- sub-emptyroom_ses-20021206_scans.tsv
     |--------- meg/
-    |------------ sub-emptyroom_ses-20021206_task-noise_channels.tsv
     |------------ sub-emptyroom_ses-20021206_task-noise_meg.json
     |------------ sub-emptyroom_ses-20021206_task-noise_meg.fif
+    |------------ sub-emptyroom_ses-20021206_task-noise_channels.tsv
     |--- sub-01/
     |------ ses-01/
     |--------- sub-01_ses-01_scans.tsv
-    |--------- meg/
-    |------------ sub-01_ses-01_task-audiovisual_run-01_events.tsv
-    |------------ sub-01_ses-01_task-audiovisual_run-01_meg.fif
-    |------------ sub-01_ses-01_task-audiovisual_run-01_meg.json
-    |------------ sub-01_ses-01_coordsystem.json
-    |------------ sub-01_ses-01_task-audiovisual_run-01_channels.tsv
     |--------- anat/
     |------------ sub-01_ses-01_T1w.json
     |------------ sub-01_ses-01_T1w.nii.gz
+    |--------- meg/
+    |------------ sub-01_ses-01_task-audiovisual_run-01_meg.json
+    |------------ sub-01_ses-01_task-audiovisual_run-01_events.tsv
+    |------------ sub-01_ses-01_task-audiovisual_run-01_channels.tsv
+    |------------ sub-01_ses-01_task-audiovisual_run-01_meg.fif
+    |------------ sub-01_ses-01_coordsystem.json
 
 
 
@@ -499,7 +499,7 @@ The data is already in a convenient form to create epochs and evokeds.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  6.539 seconds)
+   **Total running time of the script:** ( 0 minutes  6.238 seconds)
 
 
 .. _sphx_glr_download_auto_examples_convert_mne_sample.py:
