+++ _sources/auto_examples/convert_eeg_to_bids.rst.txt
@@ -133,8 +133,8 @@ of the directory tree.
     |--------------- S001R04.edf
     |------------ S002/
     |--------------- S002R12.edf
-    |--------------- S002R02.edf
     |--------------- S002R04.edf
+    |--------------- S002R02.edf
 
 
 
@@ -460,37 +460,37 @@ What does our fresh BIDS directory look like?
 
     |eegmmidb_bids/
     |--- participants.tsv
-    |--- dataset_description.json
     |--- participants.json
+    |--- dataset_description.json
     |--- sub-002/
     |------ sub-002_scans.tsv
     |------ eeg/
-    |--------- sub-002_task-imaginefists_run-1_channels.tsv
-    |--------- sub-002_task-imaginefists_run-2_eeg.edf
-    |--------- sub-002_task-imaginefists_run-2_events.tsv
     |--------- sub-002_task-resteyesclosed_eeg.json
-    |--------- sub-002_task-imaginefists_run-1_eeg.json
+    |--------- sub-002_task-imaginefists_run-1_channels.tsv
+    |--------- sub-002_task-resteyesclosed_channels.tsv
+    |--------- sub-002_task-imaginefists_run-1_events.tsv
     |--------- sub-002_task-imaginefists_run-2_eeg.json
     |--------- sub-002_task-imaginefists_run-1_eeg.edf
-    |--------- sub-002_task-imaginefists_run-1_events.tsv
-    |--------- sub-002_task-resteyesclosed_eeg.edf
-    |--------- sub-002_task-imaginefists_run-2_channels.tsv
-    |--------- sub-002_task-resteyesclosed_channels.tsv
     |--------- sub-002_task-resteyesclosed_events.tsv
+    |--------- sub-002_task-imaginefists_run-2_eeg.edf
+    |--------- sub-002_task-imaginefists_run-2_events.tsv
+    |--------- sub-002_task-imaginefists_run-2_channels.tsv
+    |--------- sub-002_task-imaginefists_run-1_eeg.json
+    |--------- sub-002_task-resteyesclosed_eeg.edf
     |--- sub-001/
     |------ sub-001_scans.tsv
     |------ eeg/
-    |--------- sub-001_task-imaginefists_run-2_eeg.json
-    |--------- sub-001_task-imaginefists_run-1_eeg.edf
-    |--------- sub-001_task-imaginefists_run-2_channels.tsv
-    |--------- sub-001_task-imaginefists_run-2_events.tsv
-    |--------- sub-001_task-resteyesclosed_eeg.json
-    |--------- sub-001_task-imaginefists_run-1_channels.tsv
-    |--------- sub-001_task-imaginefists_run-1_eeg.json
-    |--------- sub-001_task-resteyesclosed_events.tsv
-    |--------- sub-001_task-resteyesclosed_eeg.edf
-    |--------- sub-001_task-resteyesclosed_channels.tsv
     |--------- sub-001_task-imaginefists_run-1_events.tsv
+    |--------- sub-001_task-imaginefists_run-2_events.tsv
+    |--------- sub-001_task-imaginefists_run-2_channels.tsv
+    |--------- sub-001_task-resteyesclosed_events.tsv
+    |--------- sub-001_task-imaginefists_run-1_eeg.json
+    |--------- sub-001_task-resteyesclosed_eeg.json
+    |--------- sub-001_task-imaginefists_run-2_eeg.json
+    |--------- sub-001_task-resteyesclosed_eeg.edf
+    |--------- sub-001_task-imaginefists_run-1_eeg.edf
+    |--------- sub-001_task-resteyesclosed_channels.tsv
[…]
