### 2019-12-07 9a9f380
clarify intro
