### 2019-12-07 4a878df
Apply suggestions from code review
Co-Authored-By: Mainak Jas <jasmainak@users.noreply.github.com>
