### 2019-12-07 mne-tools/mne-bids#315 @sappelhoff
Thanks @jdkent

I think we could use the CoC of MNE-Python (see [here](https://github.com/mne-tools/mne-python/blob/master/CODE_OF_CONDUCT.md)) ... WDYT @jasmainak @agramfort ?

Perhaps we can do it using the GitHub organization feature? That way, a single CoC file would be shared for all MNE repositories. cc @larsoner

> You can add default community health files for your organization to the root of a public repository called .github that is owned by the organization. GitHub will use and display default files for any public repository in the organization that does not have its own file of that type in any of the following places ...

see here:

https://help.github.com/en/github/building-a-strong-community/creating-a-default-community-health-file-for-your-organization
