### 2019-12-07 mne-tools/mne-bids#316 @sappelhoff
Yes, I think it would be useful to have at least a minimal template. I'll make a draft based on your shared examples @jdkent, thanks!
