### 2019-12-07 mne-tools/mne-bids#315 @agramfort
Anything is ok for me
