+++ binder/requirements.txt
@@ -1,2 +1,0 @@
-# requirements for binder
-mne_bids~=0.3
