+++ _sources/auto_examples/convert_group_studies.rst.txt
@@ -1185,38 +1185,38 @@ Now let's see the structure of the BIDS folder we created.
  .. code-block:: none
 
     |ds000117-bids/
-    |--- participants.tsv
-    |--- participants.json
     |--- dataset_description.json
+    |--- participants.json
+    |--- participants.tsv
     |--- sub-1/
     |------ ses-01/
     |--------- sub-1_ses-01_scans.tsv
     |--------- meg/
-    |------------ sub-1_ses-01_task-VisualFaces_run-2_channels.tsv
-    |------------ sub-1_ses-01_task-VisualFaces_run-1_channels.tsv
-    |------------ sub-1_ses-01_task-VisualFaces_run-6_meg.json
-    |------------ sub-1_ses-01_task-VisualFaces_run-3_channels.tsv
-    |------------ sub-1_ses-01_task-VisualFaces_run-5_events.tsv
-    |------------ sub-1_ses-01_coordsystem.json
-    |------------ sub-1_ses-01_task-VisualFaces_run-4_events.tsv
     |------------ sub-1_ses-01_task-VisualFaces_run-3_meg.json
+    |------------ sub-1_ses-01_task-VisualFaces_run-5_meg.json
+    |------------ sub-1_ses-01_task-VisualFaces_run-1_meg.json
     |------------ sub-1_ses-01_task-VisualFaces_run-4_meg.json
     |------------ sub-1_ses-01_task-VisualFaces_run-5_channels.tsv
     |------------ sub-1_ses-01_task-VisualFaces_run-6_channels.tsv
-    |------------ sub-1_ses-01_task-VisualFaces_run-2_meg.json
-    |------------ sub-1_ses-01_task-VisualFaces_run-4_meg.fif
     |------------ sub-1_ses-01_task-VisualFaces_run-1_meg.fif
-    |------------ sub-1_ses-01_task-VisualFaces_run-6_meg.fif
-    |------------ sub-1_ses-01_task-VisualFaces_run-2_events.tsv
     |------------ sub-1_ses-01_task-VisualFaces_run-3_events.tsv
-    |------------ sub-1_ses-01_task-VisualFaces_run-5_meg.json
-    |------------ sub-1_ses-01_task-VisualFaces_run-6_events.tsv
-    |------------ sub-1_ses-01_task-VisualFaces_run-2_meg.fif
-    |------------ sub-1_ses-01_task-VisualFaces_run-1_events.tsv
-    |------------ sub-1_ses-01_task-VisualFaces_run-3_meg.fif
-    |------------ sub-1_ses-01_task-VisualFaces_run-4_channels.tsv
-    |------------ sub-1_ses-01_task-VisualFaces_run-1_meg.json
+    |------------ sub-1_ses-01_task-VisualFaces_run-5_events.tsv
+    |------------ sub-1_ses-01_task-VisualFaces_run-4_meg.fif
     |------------ sub-1_ses-01_task-VisualFaces_run-5_meg.fif
+    |------------ sub-1_ses-01_coordsystem.json
+    |------------ sub-1_ses-01_task-VisualFaces_run-2_events.tsv
+    |------------ sub-1_ses-01_task-VisualFaces_run-3_meg.fif
+    |------------ sub-1_ses-01_task-VisualFaces_run-2_meg.json
+    |------------ sub-1_ses-01_task-VisualFaces_run-1_events.tsv
+    |------------ sub-1_ses-01_task-VisualFaces_run-2_meg.fif
+    |------------ sub-1_ses-01_task-VisualFaces_run-2_channels.tsv
+    |------------ sub-1_ses-01_task-VisualFaces_run-4_channels.tsv
[…]
