+++ _sources/auto_examples/rename_brainvision_files.rst.txt
@@ -174,7 +174,7 @@ For that, see the `BrainVision Validator <bv-validator_>`_
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.020 seconds)
+   **Total running time of the script:** ( 0 minutes  0.023 seconds)
 
 
 .. _sphx_glr_download_auto_examples_rename_brainvision_files.py:
