+++ binder/environment.yml
@@ -11,6 +11,7 @@ dependencies:
 - scikit-learn
 - matplotlib
 - pip:
+  - mne_bids~=0.3
   - mne~=0.19.0
   - nibabel
   - nilearn

[…]
