### 2019-12-07 mne-tools/mne-bids#315 @larsoner
I like the organization-wide one. I think we should check that the MNE-CPP folks are okay with ours. Can you open an issue on their GitHub repo to ask?
