### 2019-12-07 1749b36
MRG: [JOSS] address remaining project review (#312)
* describe BIDS better in README

* add info on binder

* add explicit order to examples

* change emphasis

* fix docstrs in copyfiles

* consistent capitalization and punctuation in copyfiles

* fix make_bids_basename docstr

* string -> str

* fix warning in example

* remove confusing info from example

* separate printing dir tree from processing output

* Apply suggestions from code review

Co-Authored-By: Mainak Jas <jasmainak@users.noreply.github.com>

* clarify intro

* fix heading
