+++ examples/convert_eeg_to_bids.py
@@ -1,7 +1,7 @@
 """
-===============================
-Convert EEG data to BIDS format
-===============================
+===================================
+03. Convert EEG data to BIDS format
+===================================
 
 In this example, we use MNE-BIDS to create a BIDS-compatible directory of EEG
 data. Specifically, we will follow these steps:
