+++ mne_bids/copyfiles.py
@@ -41,17 +41,16 @@ def _get_brainvision_encoding(vhdr_file, verbose=False):
     Parameters
     ----------
     vhdr_file : str
-        path to the header file
+        Path to the header file.
     verbose : Bool
-        determine whether results should be logged.
-        (default False)
+        Determine whether results should be logged (default False).
 
     Returns
     -------
     enc : str
-        encoding of the .vhdr file to pass it on to open() function
+        Encoding of the .vhdr file to pass it on to open() function
         either 'UTF-8' (default) or whatever encoding scheme is specified
-        in the header
+        in the header.
 
     """
     with open(vhdr_file, 'rb') as ef:
@@ -75,13 +74,13 @@ def _get_brainvision_paths(vhdr_path):
     Parameters
     ----------
     vhdr_path : str
-        path to the header file
+        Path to the header file.
 
     Returns
     -------
     paths : tuple
-        paths to the .eeg file at index 0 and the .vmrk file
-        at index 1 of the returned tuple
+        Paths to the .eeg file at index 0 and the .vmrk file at index 1 of
+        the returned tuple.
 
     """
     fname, ext = _parse_ext(vhdr_path)
@@ -136,9 +135,9 @@ def copyfile_ctf(src, dest):
     Parameters
     ----------
     src : str
-        path to the source raw .ds folder
+        Path to the source raw .ds folder.
     dest : str
-        path to the destination of the new bids folder.
+        Path to the destination of the new bids folder.
 
     """
     _copytree(src, dest)
@@ -157,13 +156,17 @@ def copyfile_ctf(src, dest):
 def copyfile_brainvision(vhdr_src, vhdr_dest, verbose=False):
     """Copy a BrainVision file triplet to a new location and repair links.
 
+    The BrainVision file format consists of three files: .vhdr, .eeg, and .vmrk
+    The .eeg and .vmrk files associated with the .vhdr file will be given names
+    as in `vhdr_dest` with adjusted extensions. Internal file pointers will be
+    fixed.
+
     Parameters
     ----------
-    vhdr_src, vhdr_dest: str
-        The src path of the .vhdr file to be copied and the destination
-        path. The .eeg and .vmrk files associated with the .vhdr file
-        will be given names as in vhdr_dest with adjusted extensions.
-        Internal file pointers will be fixed.
+    vhdr_src : str
+        The src path of the .vhdr file to be copied.
+    vhdr_dest : str
+        The destination path of the .vhdr file.
 
     """
     # Get extenstion of the brainvision file
@@ -220,10 +223,12 @@ def copyfile_eeglab(src, dest):
     move it to an appropriate location as well as update an internal pointer
     within the .set file.
 
-    Notes
-    -----
-    Work in progress. This function will abort upon the encounter of a .fdt
-    file.
+    Parameters
+    ----------
+    src : str
+        Path to the source raw .set file.
+    dest : str
+        Path to the destination of the new .set file.
 
     """
     # Get extenstion of the EEGLAB file
@@ -265,7 +270,16 @@ def copyfile_eeglab(src, dest):
 
 
 def copyfile_bti(raw, dest):
-    """Copy BTi data."""
+    """Copy BTi data.
+
+    Parameters
+    ----------
+    raw : instance of Raw
+        An MNE-Python raw object of BTi data.
+    dest : str
+        Destination to copy the BTi data to.
+
+    """
     pdf_fname = 'c,rfDC'
     if raw.info['highpass'] is not None:
         pdf_fname = 'c,rf%0.1fHz' % raw.info['highpass']
