+++ examples/README.txt
@@ -3,7 +3,11 @@
 Examples Gallery
 ================
 
-Here we show some exemplary use cases for MNE-BIDS.
+Here we show some exemplary use cases for MNE-BIDS. Each example has a badge
+called "Binder" at the end. By clicking on that badge, you can interactively
+run the examples in your browser.
+
+The examples are loosely ordered from more general to more advanced use cases.
 
 .. contents:: Contents
    :local:
