### 2019-12-07 mne-tools/mne-bids#295 @sappelhoff
This should now be fixed: https://github.com/mne-tools/mne-bids/commit/acd5717a8c3d6c875693e53c8619d77e4f04906e

Please reopen if something is still wrong. :-)
