+++ doc/environment.yml
@@ -1,0 +1,29 @@
+name: mne_bids
+channels:
+- defaults
+- conda-forge
+dependencies:
+- python<3.7
+- pip=19
+- mkl
+- numpy>=1.14
+- scipy>=0.18.1
+- scikit-learn
+- matplotlib
+- pip:
+  - mne_bids~=0.3
+  - mne>=0.19.1
+  - nibabel
+  - pybv
+  - nilearn
+  - flake8
+  - pytest
+  - pytest-cov
+  - pytest-sugar
+  - pytest-faulthandler
+  - check-manifest
+  - sphinx
+  - pillow
+  - numpydoc
+  - sphinx_bootstrap_theme
+  - sphinx_gallery
