### 2019-12-07 66ede50
consistent capitalization, fix typos (#313)
