+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,35 @@
+---
+name: Bug report
+about: Tell us about broken, incorrect, or inconsistent behavior.
+
+---
+
+#### Describe the bug
+〈Replace this text with a clear and concise description of the bug.〉
+
+
+#### Steps to reproduce
+〈Replace this text with a code snippet or minimal working example [MWE] to
+replicate your problem, using one of the [datasets shipped with MNE-Python],
+preferably the one called [sample].
+If you can't replicate on a built-in dataset, provide also
+a link to a small, anonymized portion of your data that does yield the error.〉
+
+[MWE]: https://en.wikipedia.org/wiki/Minimal_Working_Example
+[built-in datasets]: https://mne.tools/dev/overview/datasets_index.html
+[sample]: https://mne.tools/dev/overview/datasets_index.html#sample
+
+
+#### Expected results
+〈Replace this text with a clear and concise description of what you expected to
+happen.〉
+
+
+#### Actual results
+〈Replace this text with the actual output, traceback, screenshot, or other
+description of the results.〉
+
+
+#### Additional information
+〈Replace this text with information about your system. For example through
+using MNE-Python and running `mne.sys_info()` and pasting the output here.〉
