+++ mne_bids/tests/test_write.py
@@ -30,9 +30,9 @@ from mne.io.constants import FIFF
 from mne.io.kit.kit import get_kit_info
 
 from mne_bids import (write_raw_bids, read_raw_bids, make_bids_basename,
-                      make_bids_folders, write_anat,
-                      get_anonymization_daysback)
-from mne_bids.write import _stamp_to_dt, _get_anonymization_daysback
+                      make_bids_folders, write_anat)
+from mne_bids.write import (_stamp_to_dt, _get_anonymization_daysback,
+                            get_anonymization_daysback)
 from mne_bids.tsv_handler import _from_tsv, _to_tsv
 from mne_bids.utils import _find_matching_sidecar
 from mne_bids.pick import coil_type
