+++ mne_bids/tests/test_write.py
@@ -368,8 +368,10 @@ def test_kit(_bids_validate):
     hpi_fname = op.join(data_path, 'test_mrk.sqd')
     hpi_pre_fname = op.join(data_path, 'test_mrk_pre.sqd')
     hpi_post_fname = op.join(data_path, 'test_mrk_post.sqd')
-    electrode_fname = op.join(data_path, 'test_elp.txt')
-    headshape_fname = op.join(data_path, 'test_hsp.txt')
+    #electrode_fname = op.join(data_path, 'test_elp.txt')
+    #headshape_fname = op.join(data_path, 'test_hsp.txt')
+    electrode_fname = op.join(data_path, 'test_elp.elp')
+    headshape_fname = op.join(data_path, 'test_hsp.hsp')
     event_id = dict(cond=1)
 
     kit_bids_basename = bids_basename.replace('_acq-01', '')
