### 2019-12-08 4d23814
fix test for mne<=0.19.2
