### 2019-12-08 c78c41e
fix tests when pybv is not installed
