### 2019-12-08 cca696f
remove test_elp.txt, test_hsp.txt
