### 2019-12-08 d3497d4
fix pep8 + advocate make pep on travis + activate pydocstyle
