+++ mne_bids/commands/run.py
@@ -28,7 +28,7 @@ def print_help():
 
 
 def main():
-
+    """Run main command."""
     if len(sys.argv) == 1:
         print_help()
     elif ("help" in sys.argv[1] or "-h" in sys.argv[1]):
