+++ setup.cfg
@@ -1,0 +1,18 @@
+[flake8]
+exclude = __init__.py
+ignore = W504,I101,I100,I201
+
+[tool:pytest]
+addopts =
+    --showlocals --durations=20 --doctest-modules -ra --cov-report=
+    --doctest-ignore-import-errors --junit-xml=junit-results.xml
+    --ignore=doc --ignore=logo --ignore=examples --ignore=tutorials
+    --ignore=mne/gui/_*.py
+
+[pydocstyle]
+convention = pep257
+match_dir = ^(?!\.|externals|doc|tutorials|examples|logo).*$
+match = (?!tests/__init__\.py|fixes).*\.py
+add-ignore = D100,D104,D107,D413
+add-select = D214,D215,D404,D405,D406,D407,D408,D409,D410,D411
+ignore-decorators = ^(copy_.*_doc_to_|on_trait_change|cached_property|deprecated|property|.*setter).*
