+++ .travis.yml
@@ -72,7 +72,7 @@ install:
     - |
       if [ "${DEPS}" == "minimal" ]; then
           pip install $MINIMAL_DEPENDENCIES
-          pip install pyflakes pytest pytest-cov check-manifest  # for testing
+          pip install pyflakes pytest pytest-cov check-manifest pydocstyle  # for testing
       else
           conda env create -f $CONDA_ENV --name=testenv
           source activate testenv
@@ -92,8 +92,7 @@ cache:
 
 script:
     - pytest . --cov=./mne_bids --cov-report=xml --verbose --ignore mne-python
-    - check-manifest --ignore .circleci*,doc,paper,coverage.xml,.zenodo*
-    - flake8 --count mne_bids
+    - make pep
 
 after_success:
   - bash <(curl -s https://codecov.io/bash)
