### 2019-12-08 ae2e514
fix setup + makefile
