+++ setup.cfg
@@ -4,10 +4,8 @@ ignore = W504,I101,I100,I201
 
 [tool:pytest]
 addopts =
-    --showlocals --durations=20 --doctest-modules -ra --cov-report=
-    --doctest-ignore-import-errors --junit-xml=junit-results.xml
-    --ignore=doc --ignore=logo --ignore=examples --ignore=tutorials
-    --ignore=mne/gui/_*.py
+    --showlocals --durations=20 -ra --junit-xml=junit-results.xml
+    --ignore=doc
 
 [pydocstyle]
 convention = pep257
