+++ .travis.yml
@@ -72,7 +72,7 @@ install:
     - |
       if [ "${DEPS}" == "minimal" ]; then
           pip install $MINIMAL_DEPENDENCIES
-          pip install pyflakes pytest pytest-cov check-manifest pydocstyle  # for testing
+          pip install pytest pytest-cov check-manifest pydocstyle flake8  # for testing
       else
           conda env create -f $CONDA_ENV --name=testenv
           source activate testenv
