### 2019-12-08 58d89b0
implify setup.cfg
