+++ setup.cfg
@@ -9,8 +9,6 @@ addopts =
 
 [pydocstyle]
 convention = pep257
-match_dir = ^(?!\.|externals|doc|tutorials|examples|logo).*$
-match = (?!tests/__init__\.py|fixes).*\.py
+match_dir = ^(?!\.|doc|examples).*$
 add-ignore = D100,D104,D107,D413
 add-select = D214,D215,D404,D405,D406,D407,D408,D409,D410,D411
-ignore-decorators = ^(copy_.*_doc_to_|on_trait_change|cached_property|deprecated|property|.*setter).*
