### 2019-12-08 bc1b54c
fix mne_bids/__init__.py:8:56: E261 at least two spaces before inline comment
