### 2019-12-08 7aa41f0
[MRG] use travis build with minimal deps (#319)
* use travis build with minimal deps

* fix travis file

* fix?

* fix?

* fix?

* fix?

* fix test for mne<=0.19.2

* fix tests when pybv is not installed

* fix pep8 + advocate make pep on travis + activate pydocstyle

* fix setup + makefile

* fix?

* implify setup.cfg
