+++ mne_bids/tests/test_write.py
@@ -79,7 +79,6 @@ def _bids_validate():
     return _validate
 
 
-@requires_version('pybv', '0.2.0')
 def _test_anonymize(raw, bids_basename, events_fname=None, event_id=None):
     output_path = _TempDir()
     write_raw_bids(raw, bids_basename, output_path,
@@ -387,9 +386,14 @@ def test_kit(_bids_validate):
     read_raw_bids(kit_bids_basename + '_meg.sqd', output_path)
 
     # test anonymize
-    output_path = _test_anonymize(raw, kit_bids_basename,
-                                  events_fname, event_id)
-    _bids_validate(output_path)
+    if check_version('mne', '0.20'):
+        output_path = _test_anonymize(raw, kit_bids_basename,
+                                      events_fname, event_id)
+        _bids_validate(output_path)
+    else:
+        with pytest.raises(ValueError, match='MNE is too old.'):
+            output_path = _test_anonymize(raw, kit_bids_basename,
+                                          events_fname, event_id)
 
     # ensure the channels file has no STI 014 channel:
     channels_tsv = make_bids_basename(
@@ -478,18 +482,19 @@ def test_ctf(_bids_validate):
     assert op.exists(op.join(output_path, 'participants.tsv'))
 
     # test anonymize
-    raw = mne.io.read_raw_ctf(raw_fname)
-    with pytest.warns(UserWarning,
-                      match='Converting to FIF for anonymization'):
-        output_path = _test_anonymize(raw, bids_basename)
-    _bids_validate(output_path)
+    if check_version('mne', '0.20'):
+        raw = mne.io.read_raw_ctf(raw_fname)
+        with pytest.warns(UserWarning,
+                          match='Converting to FIF for anonymization'):
+            output_path = _test_anonymize(raw, bids_basename)
+        _bids_validate(output_path)
 
-    # XXX: change the next two lines once raw.set_meas_date() is
-    # available.
-    raw.info['meas_date'] = None
-    raw.anonymize()
-    with pytest.raises(ValueError, match='All measurement dates are None'):
-        get_anonymization_daysback(raw)
+        # XXX: change the next two lines once raw.set_meas_date() is
+        # available.
+        raw.info['meas_date'] = None
+        raw.anonymize()
+        with pytest.raises(ValueError, match='All measurement dates are None'):
+            get_anonymization_daysback(raw)
 
 
 def test_bti(_bids_validate):
@@ -514,13 +519,14 @@ def test_bti(_bids_validate):
         read_raw_bids(bids_basename + '_meg', output_path,
                       extra_params=dict(foo='bar'))
 
-    # test anonymize
-    raw = mne.io.read_raw_bti(raw_fname, config_fname=config_fname,
-                              head_shape_fname=headshape_fname)
-    with pytest.warns(UserWarning,
-                      match='Converting to FIF for anonymization'):
-        output_path = _test_anonymize(raw, bids_basename)
-    _bids_validate(output_path)
+    if check_version('mne', '0.20'):
+        # test anonymize
+        raw = mne.io.read_raw_bti(raw_fname, config_fname=config_fname,
+                                  head_shape_fname=headshape_fname)
+        with pytest.warns(UserWarning,
+                          match='Converting to FIF for anonymization'):
+            output_path = _test_anonymize(raw, bids_basename)
+        _bids_validate(output_path)
 
 
 # XXX: vhdr test currently passes only on MNE master. Skip until next release.
@@ -575,8 +581,10 @@ def test_vhdr(_bids_validate):
[…]
