+++ setup.cfg
@@ -1,0 +1,14 @@
+[flake8]
+exclude = __init__.py
+ignore = W504,I101,I100,I201
+
+[tool:pytest]
+addopts =
+    --showlocals --durations=20 -ra --junit-xml=junit-results.xml
+    --ignore=doc
+
+[pydocstyle]
+convention = pep257
+match_dir = ^(?!\.|doc|examples).*$
+add-ignore = D100,D104,D107,D413
+add-select = D214,D215,D404,D405,D406,D407,D408,D409,D410,D411
