+++ Makefile
@@ -26,7 +26,7 @@ clean-cache:
 clean: clean-build clean-pyc clean-so clean-ctags clean-cache
 
 inplace:
-	$(PYTHON) setup.py install
+	$(PYTHON) setup.py develop
 
 test: inplace check-manifest
 	rm -f .coverage
@@ -63,7 +63,7 @@ pydocstyle:
 	@pydocstyle
 
 pep:
-	@$(MAKE) -k flake pydocstyle
+	@$(MAKE) -k flake pydocstyle check-manifest
 
 build-doc:
 	cd doc; make clean
