+++ mne_bids/write.py
@@ -244,8 +244,12 @@ def _participants_tsv(raw, subject_id, fname, overwrite=False,
             meas_date = meas_date[0]
 
         if meas_date is not None and age is not None:
-            bday = datetime(age[0], age[1], age[2])
-            meas_datetime = datetime.fromtimestamp(meas_date)
+            bday = datetime(age[0], age[1], age[2], tzinfo=timezone.utc)
+            meas_datetime = meas_date
+            if isinstance(meas_date, datetime):
+                meas_datetime = meas_date
+            else:
+                meas_datetime = datetime.fromtimestamp(meas_date)
             subject_age = _age_on_date(bday, meas_datetime)
         else:
             subject_age = "n/a"
@@ -1120,8 +1124,9 @@ def write_raw_bids(raw, bids_basename, output_path, events_data=None,
         # check the session date provided is consistent with the value in raw
         meas_date = raw.info.get('meas_date', None)
         if meas_date is not None:
-            er_date = datetime.fromtimestamp(
-                raw.info['meas_date'][0]).strftime('%Y%m%d')
+            if not isinstance(meas_date, datetime):
+                meas_date = datetime.fromtimestamp(meas_date[0])
+            er_date = meas_date.strftime('%Y%m%d')
             if er_date != session_id:
                 raise ValueError("Date provided for session doesn't match "
                                  "session date.")
