### 2019-12-09 4b3655c
coerce nan duration to 0
