### 2019-12-09 d54edfd
implement copyfile_kit function in copyfiles.py
