### 2019-12-09 c3239d8
DOC improve contributing guide
