### 2019-12-09 mne-tools/mne-bids#315 @sappelhoff
@larsoner we got approval from the MNE-CPP folks in https://github.com/mne-tools/mne-cpp/issues/418#issuecomment-563035899.

Could you open the new `.github` repo and put the MNE-Python Code of Conduct there, please? I don't think I have the required rights for this.
