### 2019-12-09 5fd9bd3
add test for copyfile_kit
