### 2019-12-09 ebf0daf
fix flake8 errors
