### 2019-12-09 0362915
add test for sync envs
