+++ doc/environment.yml
@@ -22,6 +22,7 @@ dependencies:
   - pytest-sugar
   - pytest-faulthandler
   - check-manifest
+  - pydocstyle
   - sphinx
   - pillow
   - numpydoc
