### 2019-12-09 e0e96b4
add @jdkent's recommendations for the paper
