### 2019-12-09 mne-tools/mne-bids#315 @larsoner
Done

https://github.com/mne-tools/.github/blob/master/CODE_OF_CONDUCT.md
