### 2019-12-09 27dfa06
DOC: overhaul mne sample data example
