+++ examples/convert_mne_sample.py
@@ -5,8 +5,10 @@
 02. Convert MNE sample data to BIDS format
 ==========================================
 
-This example demonstrates how to convert your existing files into a
-BIDS-compatible folder.
+In this example we will use MNE-BIDS to organize the MNE sample data according
+to the BIDS standard.
+In a second step we will read the organized dataset using MNE-BIDS.
+
 """
 
 # Authors: Mainak Jas <mainak.jas@telecom-paristech.fr>
@@ -17,10 +19,10 @@ BIDS-compatible folder.
 # License: BSD (3-clause)
 
 ###############################################################################
-# Let us import mne_bids
+# First we import some basic Python libraries, followed by MNE-Python and its
+# sample data, and then finally the MNE-BIDS functions we need for this example
 
 import os.path as op
-
 from datetime import datetime
 
 import mne
@@ -30,7 +32,11 @@ from mne_bids import write_raw_bids, read_raw_bids, make_bids_basename
 from mne_bids.utils import print_dir_tree
 
 ###############################################################################
-# And define the paths and event_id dictionary.
+# Now we can read the MNE sample data. We define an `event_id` based on our
+# knowledge of the data, to give meaning to events in the data.
+#
+# With `raw_fname` and `events_data` we determine where to get the sample data
+# from. `output_path` determines where we will write the BIDS conversion to.
 
 data_path = sample.data_path()
 event_id = {'Auditory/Left': 1, 'Auditory/Right': 2, 'Visual/Left': 3,
@@ -44,12 +50,16 @@ output_path = op.join(data_path, '..', 'MNE-sample-data-bids')
 #
 # .. note::
 #
-#   ``mne-bids`` will infer the manufacturer information from the file
-#   extension. If your file format is non-standard for the
-#   manufacturer, please update the manufacturer field in the sidecars
-#   manually.
+#   ``mne-bids`` will try to infer as much information from the data as
+#   possible to then save this data in BIDS specific "sidecar" files. For
+#   example the manufacturer information, which is inferred from the data file
+#   extension. However, sometimes inferring is ambiguous (e.g., if your file
+#   format is non-standard for the manufacturer). In these cases, MNE-BIDS does
+#   *not* guess and you will have to update your BIDS fields manually.
 #
-# Specify the raw_file and events_data and run the BIDS conversion.
+# Based on our path definitions above, we read the raw data file, define
+# a new BIDS name for it, and then run the automatic BIDS conversion.
+
 raw = mne.io.read_raw_fif(raw_fname)
 bids_basename = make_bids_basename(subject='01', session='01',
                                    task='audiovisual', run='01')
@@ -57,11 +67,16 @@ write_raw_bids(raw, bids_basename, output_path, events_data=events_data,
                event_id=event_id, overwrite=True)
 
 ###############################################################################
-# Specify some empty room data and run BIDS conversion on it.
+# We are dealing with MEG data, which is often accompanied by so-called
+# "empty room" recordings for noise modeling. Below we show that we can use
+# MNE-BIDS to also save such a recording with the just converted data.
+
+# Specify the path to the not-yet-BIDS empty room data and read it
 er_raw_fname = op.join(data_path, 'MEG', 'sample', 'ernoise_raw.fif')
 er_raw = mne.io.read_raw_fif(er_raw_fname)
+
 # For empty room data we need to specify that the subject ID is
-# 'emptyroom', and that the task is 'noise'.
+# 'emptyroom', and that the task is 'noise' (these are BIDS rules).
 # We also need to specify the recording date in the format YYYYMMDD for the
 # session id.
 meas_date = er_raw.info['meas_date']
@@ -69,19 +84,27 @@ if not isinstance(meas_date, datetime):
[…]
