### 2019-12-09 mne-tools/mne-bids#315 @sappelhoff
awesome, thanks a lot @larsoner and @jdkent for bringing up this point.
