### 2019-12-09 mne-tools/mne-bids#304 @sappelhoff
I addressed the two last parts (on improving the examples) in #329. I added more explanations to the basic dataset conversion example ... and then in the group conversion example I recommend users to do the basic example first, so that they understand what's going on.
