### 2019-12-10 b5cefb6
Rebased and retested with fresh master branch.
