+++ mne_bids/tests/test_write.py
@@ -334,7 +334,7 @@ def test_fif(_bids_validate):
         prefix=op.join(bids_root, 'sub-01', 'ses-01'))
     data = _from_tsv(scans_tsv)
     assert data['acq_time'][0] == 'n/a'
-    _bids_validate(output_path)
+    _bids_validate(bids_root)
 
     # check that split files have part key
     raw = mne.io.read_raw_fif(raw_fname)
