### 2019-12-10 c8dd671
Fixed pep8 issues.
