### 2019-12-10 e74e866
fix flake8 errors
