### 2019-12-10 6d8be3c
fix inconsistent output_path at test
