### 2019-12-10 e89f679
Adding fixes to output_paths.
