### 2019-12-10 c9715de
Adding flake8 fixes.
