### 2019-12-10 e1673ce
Address Stefan's comments
Co-Authored-By: Stefan Appelhoff <stefan.appelhoff@mailbox.org>
