### 2019-12-10 8545743
MRG: Standardizing bids_root / output_path for the read/write API (#303)
* Adding updated bids root.

* Fixing readme to reflect correct change.

* Fixing examples to show the bids_root.

* Fixing commands in the mne_bids examples.

* Fixing additional locations of output_path renaming to bids_root.

* Update mne_bids/tests/test_write.py

Co-Authored-By: Mainak Jas <jasmainak@users.noreply.github.com>

* Refactored everything from output_path to bids_root.

* Rebased and retested with fresh master branch.

* Adding mne-python to git ignore.

* Fixed pep8 issues.

* Adding fixes to output_paths
[…]
