### 2019-12-11 a9be3ef
Adding existence check to the mne write raw bids.
