### 2019-12-11 mne-tools/mne-bids#304 @sappelhoff
Everything has been addressed :-)
