### 2019-12-11 9f7920e
MRG: binder install with distinct env yml (#322)
* binder install with distinct env yml

* add test for sync envs

* fix indentation
