### 2019-12-11 5d89955
Adding fixed updated test to check dataset description existence.
