+++ mne_bids/write.py
@@ -1164,7 +1164,8 @@ def write_raw_bids(raw, bids_basename, bids_root, events_data=None,
     if events is not None and len(events) > 0 and not emptyroom:
         _events_tsv(events, raw, events_fname, event_id, overwrite, verbose)
 
-    if not op.exists(op.join(bids_root, "dataset_description.json")) or overwrite:
+    if not op.exists(op.join(bids_root, "dataset_description.json")) \
+            or overwrite:
         make_dataset_description(bids_root, name=" ", verbose=verbose)
     _sidecar_json(raw, task, manufacturer, sidecar_fname, kind, overwrite,
                   verbose)
