### 2019-12-11 mne-tools/mne-bids#305 @sappelhoff
all points have been addressed now :-) closing.
