### 2019-12-11 #1896 @sappelhoff
Dear @jdkent and @TomDonoghue we believe that all your valuable feedback has now been addressed. We took extra care to cross link from each of your raised issues to the PRs solving them, so the progress should be thoroughly documented on GitHub.

Thanks again for your reviews - this has really improved our current status with the package!
