### 2019-12-11 #1896 @TomDonoghue
Thank you @sappelhoff for all the detailed responses and clear linking between comments and PRs - I've been able to follow along clearly with all the updates to all the points, and everything looks great to me!

I think this is a great package in great shape! I have ticked of my remaining review points, and from my end I absolutely think this project / paper is good to be published in JOSS.
