### 2019-12-11 38741c0
Adding default authors as MNE-BIDS.
