### 2019-12-11 #1896 @jdkent
Thanks @sappelhoff! All the issues I raised were followed up on and all of them have reached a satisfactory conclusion. I have happily checked off all remaining tick marks.

Another :+1: from me for publishing this project in JOSS.
