### 2019-12-11 ebb8051
Fixing test setup of dataset description write.
