### 2019-12-12 76b43e0
do a nested import make_bids_basename
