### 2019-12-12 #1896 @arokem
Yep - looks great. I added a small suggestion to language in https://github.com/mne-tools/mne-bids/pull/333.

Once that's merged or closed, could you please:

- [ ] Please list the Zenodo DOI of the archived version here.
