### 2019-12-12 a887666
Fixed to check for a list.
