### 2019-12-12 mne-tools/mne-bids#334 @sappelhoff
wrap up JOSS publication
Hi everyone,

mne-bids is now accepted for publication and ready to be released :-)

We have a final list of to dos including a release, see here: https://github.com/openjournals/joss-reviews/issues/1896#issuecomment-564831574

Can we do a release in the next days @jasmainak @agramfort ?
