+++ mne_bids/tests/test_copyfiles.py
@@ -151,10 +151,10 @@ def test_copyfile_kit():
     _, ext = _parse_ext(hpi_fname, verbose=True)
     if ext == '.sqd':
         assert op.exists(op.join(
-            output_path, kit_bids_basename+'_markers.sqd'))
+            output_path, kit_bids_basename + '_markers.sqd'))
     elif ext == '.mrk':
         assert op.exists(op.join(
-            output_path, kit_bids_basename+'_markers.mrk'))
+            output_path, kit_bids_basename + '_markers.mrk'))
 
     if op.exists(electrode_fname):
         task, run, key = None, None, 'ELP'
