+++ mne_bids/copyfiles.py
@@ -197,24 +197,16 @@ def copyfile_kit(src, dest, subject_id, session_id,
                 acquisition=key, suffix='markers%s' % marker_ext,
                 prefix=data_path)
             sh.copyfile(value, marker_fname)
-    if 'elp' in _init_kwargs:
-        hpi_file = _init_kwargs['elp']
-        task, run, acq = None, None, 'ELP'
-        marker_ext = '.pos'
-        marker_fname = make_bids_basename(
-            subject=subject_id, session=session_id, task=task, run=run,
-            acquisition=acq, suffix='headshape%s' % marker_ext,
-            prefix=data_path)
-        sh.copyfile(hpi_file, marker_fname)
-    if 'hsp' in _init_kwargs:
-        hpi_file = _init_kwargs['hsp']
-        task, run, acq = None, None, 'HSP'
-        marker_ext = '.pos'
-        marker_fname = make_bids_basename(
-            subject=subject_id, session=session_id, task=task, run=run,
-            acquisition=acq, suffix='headshape%s' % marker_ext,
-            prefix=data_path)
-        sh.copyfile(hpi_file, marker_fname)
+    for acq in ['elp', 'hsp']:
+        if acq in _init_kwargs:
+            hpi_file = _init_kwargs['elp']
+            task, run, acq = None, None, acq.upper()
+            marker_ext = '.pos'
+            marker_fname = make_bids_basename(
+                subject=subject_id, session=session_id, task=task, run=run,
+                acquisition=acq, suffix='headshape%s' % marker_ext,
+                prefix=data_path)
+            sh.copyfile(hpi_file, marker_fname)
 
 
 def copyfile_brainvision(vhdr_src, vhdr_dest, verbose=False):
