### 2019-12-12 mne-tools/mne-bids#334 @agramfort
sure no objection.

>
