### 2019-12-12 3f94850
Adding updates to incorporate stephan's changes.
