### 2019-12-12 mne-tools/mne-bids#334 @jasmainak
Okay for me. Do you do a minor release or major?
