### 2019-12-12 a33c3bc
BIDS compatible ieeg coordinates writing.
