+++ mne_bids/write.py
@@ -23,6 +23,7 @@ from mne import Epochs
 from mne.io.constants import FIFF
 from mne.io.pick import channel_type
 from mne.io import BaseRaw, anonymize_info, read_fiducials
+
 try:
     from mne.io._digitization import _get_fid_coords
 except ImportError:
@@ -342,7 +343,7 @@ def _scans_tsv(raw, raw_fname, fname, overwrite=False, verbose=True):
         acq_time = meas_date.strftime('%Y-%m-%dT%H:%M:%S')
 
     data = OrderedDict([('filename', ['%s' % raw_fname.replace(os.sep, '/')]),
-                       ('acq_time', [acq_time])])
+                        ('acq_time', [acq_time])])
 
     if os.path.exists(fname):
         orig_data = _from_tsv(fname)
@@ -359,14 +360,19 @@ def _scans_tsv(raw, raw_fname, fname, overwrite=False, verbose=True):
 
     return fname
 
-def _electrodes_tsv(raw, fname, overwrite=False, verbose=True):
-    """Create an electrodes.tsv file and save it.
+
+def _electrodes_tsv(raw, fname, kind, overwrite=False, verbose=True):
+    """
+    Create an electrodes.tsv file and save it.
+
     Parameters
     ----------
     raw : instance of Raw
         The data as MNE-Python Raw object.
     fname : str
         Filename to save the electrodes.tsv to.
+    kind : str
+        Acquisition type to save electrodes in. For iEEG, requires size.
     overwrite : bool
         Defaults to False.
         Whether to overwrite the existing data in the file.
@@ -386,13 +392,24 @@ def _electrodes_tsv(raw, fname, overwrite=False, verbose=True):
             y.append('n/a')
             z.append('n/a')
         names.append(ch['ch_name'])
-    data = OrderedDict([('name', names),
-                        ('x', x),
-                        ('y', y),
-                        ('z', z),
-                        ])
+
+    if kind == "ieeg":
+        sizes = ['n/a'] * len(names)
+        data = OrderedDict([('name', names),
+                            ('x', x),
+                            ('y', y),
+                            ('z', z),
+                            ('size', sizes),
+                            ])
+    else:
+        data = OrderedDict([('name', names),
+                            ('x', x),
+                            ('y', y),
+                            ('z', z),
+                            ])
     _write_tsv(fname, data, overwrite=overwrite, verbose=verbose)
     return fname
+
 
 def _coordsystem_json(raw, unit, orient, coordsystem_name, fname,
                       kind, overwrite=False, verbose=True):
@@ -452,12 +469,10 @@ def _coordsystem_json(raw, unit, orient, coordsystem_name, fname,
                     'AnatomicalLandmarkCoordinateUnits': unit,
                     }
     elif kind == "ieeg":
-        fid_json = {'iEEGCoordinateSystem': coordsystem_name, # Pixels, or ACPC
-                    'iEEGCoordinateUnits': unit, # m, mm, cm , or pixels
-                    'AnatomicalLandmarkCoordinates': coords,
-                    'AnatomicalLandmarkCoordinateSystem': coordsystem_name,
-                    'AnatomicalLandmarkCoordinateUnits': unit,
-                    }
+        fid_json = {
+            'iEEGCoordinateSystem': coordsystem_name,  # Pixels, or ACPC
+            'iEEGCoordinateUnits': unit,  # m, mm, cm , or pixels
+        }
     else:
         warn('Writing of electrodes.tsv is not supported for kind "{}". '
[…]
