### 2019-12-12 4ffa26e
FIX travis by using npm2
