### 2019-12-13 mne-tools/mne-bids#334 @sappelhoff
> Do you do a minor release or major?

I think any release would be fine, as long as it gets stored by zenodo. I have already enabled zenodo for this project ... so the next release (minor or major) will be stored.

I would be fine with either type of release.
