### 2019-12-13 mne-tools/mne-bids#334 @jasmainak
fair enough!
