+++ mne_bids/copyfiles.py
@@ -199,14 +199,14 @@ def copyfile_kit(src, dest, subject_id, session_id,
             sh.copyfile(value, marker_fname)
     for acq in ['elp', 'hsp']:
         if acq in _init_kwargs:
-            hpi_file = _init_kwargs['elp']
+            position_file = _init_kwargs[acq]
             task, run, acq = None, None, acq.upper()
-            marker_ext = '.pos'
-            marker_fname = make_bids_basename(
+            position_ext = '.pos'
+            position_fname = make_bids_basename(
                 subject=subject_id, session=session_id, task=task, run=run,
-                acquisition=acq, suffix='headshape%s' % marker_ext,
+                acquisition=acq, suffix='headshape%s' % position_ext,
                 prefix=data_path)
-            sh.copyfile(hpi_file, marker_fname)
+            sh.copyfile(position_file, position_fname)
 
 
 def copyfile_brainvision(vhdr_src, vhdr_dest, verbose=False):
