### 2019-12-15 7ef57eb
Include contribution guide via m2r
