+++ doc/contribute.rst
@@ -1,0 +1,5 @@
+:orphan:
+
+.. _contribute:
+
+.. mdinclude:: ../CONTRIBUTING.md
