### 2019-12-15 3a6f914
DOC: Add information on updating conda package
