### 2019-12-15 0b93038
Add m2r dependency to conda env
