### 2019-12-16 mne-tools/mne-bids#334 @arokem
Just to confirm: whatever suits you folks, as long as there is a Zenodo DOI associated with the version as accepted.
