### 2019-12-16 mne-tools/mne-bids#334 @sappelhoff
I that case, I suggest to do a GitHub-only release (`0.3.1`), to trigger Zenodo. I think we're not ready for a big release (`0.4`) yet. And the GitHub-only `0.3.1` release will store our repo with no repercussions.

@jasmainak @agramfort - I'll do it once

1. CIs are fixed
1. I have the OK from both of you.
