+++ .travis.yml
@@ -43,7 +43,7 @@ before_install:
   - export PATH=${HOME}/miniconda/bin:$PATH
   - conda update --yes --quiet conda
   - npm install -g npm stable
-  - npm install -g node@10.16.1
+  - npm install -g node@10.16.1 --force
   - npm --version
   - node --version
   - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then
