### 2019-12-17 mne-tools/mne-bids#334 @jasmainak
yep I prefer a quick minor release. And okay for me to go ahead and make it!
