### 2019-12-17 mne-tools/mne-bids#334 @sappelhoff
done :-) the DOI should work within the next minutes.

[![DOI](https://zenodo.org/badge/89170358.svg)](https://zenodo.org/badge/latestdoi/89170358)
