### 2019-12-17 #1896 @sappelhoff
Hi @arokem we made a release and archived the software with Zenodo:

[![DOI](https://zenodo.org/badge/89170358.svg)](https://zenodo.org/badge/latestdoi/89170358)

The metadata is aligned with the paper (including title and authors ... orcids are added as well).

Thanks also for your language suggestion, which has been incorporated :+1:
