### 2019-12-17 1321199
update whats_new.rst
