### 2019-12-17 308b300
update whats_new.rst
