### 2019-12-17 541d82f
update whats_new.rst
