### 2019-12-17 ed5192a
update whats_new.rst
