### 2019-12-17 5a59033
update whats_new.rst
