### 2019-12-17 451fd9e
update whats_new.rst
